n2o4::cfe::es
Function run_loop

pub fn run_loop(run_status: Option<RunStatus>) -> bool

Checks for exit requests from the cFE system and possibly makes a request for app shutdown to the cFE system.

If run_status is set to Some(AppExit) or Some(AppError), the cFE system treats the function call as a shutdown request for this application.

Returns whether the app should continue running; a return value of false means the application should gracefully shut down.

Wraps CFE_ES_RunLoop.