n2o4::cfe::es

Function write_to_syslog21

pub fn write_to_syslog21<A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U>(
    fmt: PrintfFmt<(A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U)>,
    a: A,
    b: B,
    c: C,
    d: D,
    e: E,
    f: F,
    g: G,
    h: H,
    i: I,
    j: J,
    k: K,
    l: L,
    m: M,
    n: N,
    o: O,
    p: P,
    q: Q,
    r: R,
    s: S,
    t: T,
    u: U,
) -> Status
where
    A: PrintfArgument,
    B: PrintfArgument,
    C: PrintfArgument,
    D: PrintfArgument,
    E: PrintfArgument,
    F: PrintfArgument,
    G: PrintfArgument,
    H: PrintfArgument,
    I: PrintfArgument,
    J: PrintfArgument,
    K: PrintfArgument,
    L: PrintfArgument,
    M: PrintfArgument,
    N: PrintfArgument,
    O: PrintfArgument,
    P: PrintfArgument,
    Q: PrintfArgument,
    R: PrintfArgument,
    S: PrintfArgument,
    T: PrintfArgument,
    U: PrintfArgument,

Writes a message to the cFE System Log using a format string and 21 format arguments.

Wraps CFE_ES_WriteToSysLog.