n2o4::cfe::evs
Struct EventSender

pub struct EventSender { /* private fields */ }

A marker type to ensure you register before sending events.

Implementations

impl EventSender

pub fn send_event0( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<()>, ) -> Status

Generates a software event using a format string and 0 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id0( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<()>, ) -> Status

Generates a software event (with the specified Application ID) using a format string and 0 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event0( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<()>, ) -> Status

Generates a software event (with a specific time tag) using a format string and 0 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event1<A>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A,)>, a: A, ) -> Status
where A: PrintfArgument,

Generates a software event using a format string and 1 format argument.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id1<A>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A,)>, a: A, ) -> Status
where A: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 1 format argument.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event1<A>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A,)>, a: A, ) -> Status
where A: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 1 format argument.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event2<A, B>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B)>, a: A, b: B, ) -> Status
where A: PrintfArgument, B: PrintfArgument,

Generates a software event using a format string and 2 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id2<A, B>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B)>, a: A, b: B, ) -> Status
where A: PrintfArgument, B: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 2 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event2<A, B>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B)>, a: A, b: B, ) -> Status
where A: PrintfArgument, B: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 2 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event3<A, B, C>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C)>, a: A, b: B, c: C, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument,

Generates a software event using a format string and 3 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id3<A, B, C>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C)>, a: A, b: B, c: C, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 3 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event3<A, B, C>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C)>, a: A, b: B, c: C, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 3 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event4<A, B, C, D>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D)>, a: A, b: B, c: C, d: D, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument,

Generates a software event using a format string and 4 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id4<A, B, C, D>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C, D)>, a: A, b: B, c: C, d: D, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 4 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event4<A, B, C, D>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D)>, a: A, b: B, c: C, d: D, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 4 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event5<A, B, C, D, E>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E)>, a: A, b: B, c: C, d: D, e: E, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument,

Generates a software event using a format string and 5 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id5<A, B, C, D, E>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C, D, E)>, a: A, b: B, c: C, d: D, e: E, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 5 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event5<A, B, C, D, E>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E)>, a: A, b: B, c: C, d: D, e: E, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 5 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event6<A, B, C, D, E, F>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F)>, a: A, b: B, c: C, d: D, e: E, f: F, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument,

Generates a software event using a format string and 6 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id6<A, B, C, D, E, F>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C, D, E, F)>, a: A, b: B, c: C, d: D, e: E, f: F, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 6 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event6<A, B, C, D, E, F>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F)>, a: A, b: B, c: C, d: D, e: E, f: F, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 6 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event7<A, B, C, D, E, F, G>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F, G)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument,

Generates a software event using a format string and 7 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id7<A, B, C, D, E, F, G>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C, D, E, F, G)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 7 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event7<A, B, C, D, E, F, G>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F, G)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 7 format arguments.

Wraps CFE_EVS_SendTimedEvent.

pub fn send_event8<A, B, C, D, E, F, G, H>( &self, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F, G, H)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument, H: PrintfArgument,

Generates a software event using a format string and 8 format arguments.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id8<A, B, C, D, E, F, G, H>( &self, event_id: u16, event_type: EventType, app_id: AppId, fmt: PrintfFmt<(A, B, C, D, E, F, G, H)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument, H: PrintfArgument,

Generates a software event (with the specified Application ID) using a format string and 8 format arguments.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event8<A, B, C, D, E, F, G, H>( &self, time: SysTime, event_id: u16, event_type: EventType, fmt: PrintfFmt<(A, B, C, D, E, F, G, H)>, a: A, b: B, c: C, d: D, e: E, f: F, g: G, h: H, ) -> Status
where A: PrintfArgument, B: PrintfArgument, C: PrintfArgument, D: PrintfArgument, E: PrintfArgument, F: PrintfArgument, G: PrintfArgument, H: PrintfArgument,

Generates a software event (with a specific time tag) using a format string and 8 format arguments.

Wraps CFE_EVS_SendTimedEvent.

impl EventSender

pub fn send_event_str( &self, event_id: u16, event_type: EventType, msg: &str, ) -> Status

Generates a software event using a str as the message.

Note that any embedded null characters and anything past them will not get put into the event message.

Wraps CFE_EVS_SendEvent.

pub fn send_event_with_app_id_str( &self, event_id: u16, event_type: EventType, app_id: AppId, msg: &str, ) -> Status

Generates a software event with the specified Application ID using a str as the message.

Note that any embedded null characters and anything past them will not get put into the event message.

Wraps CFE_EVS_SendEventWithAppID.

pub fn send_timed_event_str( &self, time: SysTime, event_id: u16, event_type: EventType, msg: &str, ) -> Status

Generates a software event with a specific time tag using a str as the message.

Note that any embedded null characters and anything past them will not get put into the event message.

Wraps CFE_EVS_SendTimedEvent.