n2o4::cfe::evs

Trait FilterScheme

pub trait FilterScheme: FilterSchemeSealed {
    const SCHEME_ID: u16;
}

A scheme for filtering event messages so that not all get recorded.

This is a sealed trait: cFE only supports a fixed set of filter schemes.

Required Associated Constants

const SCHEME_ID: u16

An integer identifying the scheme in question.

This gets used as the FilterScheme argument to CFE_EVS_Register.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.