Struct n2o4::cfe::msg::Telemetry

#[repr(C)]
pub struct Telemetry<T: Copy> {
    pub payload: T,
    /* private fields */
}

A telemetry message for use with the cFE software bus.

Wraps CFE_MSG_TelemetryHeader_t, with a user-specified payload following.

Fields
payload: T

The message’s payload. As messages are copied willy-nilly by cFE, T needs to be Copy.

Implementations

impl<T: Copy + Sized> Telemetry<T>

pub fn new(msg_id: MsgId, payload: T) -> Result<Self, Status>

Tries to create a new telemetry message, setting the message ID along the way.

Wraps CFE_MSG_Init and CFE_MSG_GetTypeFromMsgId.

impl<T: Copy + Sized + Default> Telemetry<T>

pub fn new_default(msg_id: MsgId) -> Result<Self, Status>

new using T::default() as the payload.

impl<T: Copy + Sized, const SIZE: usize> Telemetry<[T; SIZE]>

pub fn transmit_partial( &mut self, increment_sequence_count: bool, len: usize ) -> Result<(), Status>

Transmits onto the software bus the header and first min(len, SIZE) elements of payload.

After transmission, attempts to set the message size (in the header) back to the structure’s full length.

Wraps CFE_MSG_SetSize and CFE_SB_TransmitMsg.