n2o4::cfe::sb
Struct Pipe

pub struct Pipe { /* private fields */ }

A software bus pipe; an application needs one of these to receive messages.

This may not be used from a different thread from the one it was created on.

Wraps CFE_SB_PipeId_t.

Implementations

impl Pipe

pub fn new<S: AsRef<CStr> + ?Sized>( depth: u16, pipe_name: &S, ) -> Result<Pipe, Status>

Creates a new pipe with space for depth yet-to-be-handled messages and the name pipe_name.

Wraps CFE_SB_CreatePipe.

pub fn delete(self)

Deletes the software bus pipe.

Note that applications should not call this if the deletion is part of application shutdown; the framework will do the needed cleanup at application exit.

Wraps CFE_SB_DeletePipe.

pub fn subscribe(&mut self, msg_id: MsgId) -> Result<(), Status>

Subscribes to messages with ID msg_id on the software bus with default parameters.

Wraps CFE_SB_Subscribe.

pub fn subscribe_ex( &mut self, msg_id: MsgId, quality: Qos, msg_lim: u16, ) -> Result<(), Status>

Subscribes to messages with ID msg_id on the software bus with the specified quality of service (currently unused by cFE) and a limit to the number of messages with this ID allowed in the pipe at the same time.

Wraps CFE_SB_SubscribeEx.

pub fn subscribe_local( &mut self, msg_id: MsgId, msg_lim: u16, ) -> Result<(), Status>

Subscribes to messages with ID msg_id, but keep the subscription local to the current CPU.

This is typically only used by the SBN application.

Wraps CFE_SB_SubscribeLocal.

pub fn unsubscribe(&mut self, msg_id: MsgId) -> Result<(), Status>

Removes the current pipe’s subscription to messages with ID msg_id.

Wraps CFE_SB_Unsubscribe.

pub fn unsubscribe_local(&mut self, msg_id: MsgId) -> Result<(), Status>

Removes the current pipe’s subscription to messages with ID msg_id, keeping the unsubscription local to the current CPU.

This is typically only used by the SBN application.

Wraps CFE_SB_UnsubscribeLocal.

pub fn receive_buffer<T, F>(&mut self, time_out: TimeOut, closure: F) -> T
where F: for<'a> FnOnce(Result<&'a Message, Status>) -> T,

Receives a message from the pipe.

Whether or not a message was received, closure gets called with the result of the reception attempt.

Uses time_out to determine how long to wait for a message if the pipe is empty.

Passing the message buffer to a closure rather than returning it ensures that the buffer’s lifetime constraints are respected.

Wraps CFE_SB_ReceiveBuffer.