n2o4::cfe::tbl
Struct DumpOnlyTblHandle

pub struct DumpOnlyTblHandle<T: TableType> { /* private fields */ }

A handle to a dump-only table.

Wraps a CFE_TBL_Handle_t.

Safety and Concurrency

As the only writer to the table will be the owner of the DumpOnlyTblHandle, this is safe for the owner to use. Other applications with a handle to the table may have concurrency-related problems reading data unless care is taken; that care is not automatically provided in full by DumpOnlyTblHandle. 1

There is some—non-comprehensive!—assistance in get_mut. ↩

Implementations

impl<T: TableType> DumpOnlyTblHandle<T>

pub fn register<S: AsRef<CStr> + ?Sized>( tbl_name: &S, tbl_buffer: Option<&'static mut T>, validation_fn: Option<TableValidationFn<T>>, ) -> Result<Self, Status>

Tries to register a dump-only table (with optional user-defined address tbl_buffer) with cFE, returning a handle if successful.

Wraps CFE_TBL_Register (and for tables with a user-defined address, CFE_TBL_Load).

pub fn get_mut<F, V>(&mut self, closure: F) -> Result<V, Status>
where F: for<'a> FnOnce(&'a mut T) -> V,

Attempts to obtain the current address of the table contents. If successful, provides closure with a mutable reference to the data backing the table, returning closure’s return value.

Calls CFE_TBL_Modified after closure finishes to let Table Services know the table has been modified.

In the case when the table doesn’t have a user-defined address, also wraps CFE_TBL_GetAddress and CFE_TBL_ReleaseAddress.

pub fn unregister(self) -> Result<(), Status>

Unregisters the table corresponding to this handle.

Note that you generally shouldn’t need to call this, as cFE automatically unregisters all tables owned by an application when that application exits.

In the case of a table with user-defined address, this consciously does not return the &'static mut T to the backing data, as other applications may still have handles to the table.

Wraps CFE_TBL_Unregister.