n2o4::cfe::tbl
Struct SharedTblHandle

pub struct SharedTblHandle<T: TableType> { /* private fields */ }

A handle to a table registered by another application.

Wraps a CFE_TBL_Handle_t obtained from CFE_TBL_Share.

Implementations

impl<T: TableType> SharedTblHandle<T>

pub unsafe fn share<S: AsRef<CStr> + ?Sized>( tbl_name: &S, ) -> Result<Self, Status>

Tries to obtain a handle to a table registered by another application.

Wraps CFE_TBL_Share.

Safety

The cFE API provides no way to ensure that the table named tbl_name is, in fact, a value of type T. This function just assumes that it is; this fact must be verified by the programmer.