Struct n2o4::cfe::tbl::TblHandle

pub struct TblHandle<T: TableType> { /* private fields */ }

A handle to a table.

Wraps a CFE_TBL_Handle_t.

Implementations

impl<T: TableType> TblHandle<T>

pub fn register<S: AsRef<CStr> + ?Sized>( tbl_name: &S, options: TblOptions, validation_fn: Option<TableValidationFn<T>> ) -> Result<(Self, RegisterInfo), Status>

Tries to register a loadable table with cFE, returning a handle if successful.

Wraps CFE_TBL_Register.

pub fn get_ref<F, V>(&mut self, closure: F) -> Result<V, Status>where F: for<'a> FnOnce(&'a T, bool) -> V,

Tries to obtain the current address of the table contents. If successful, passes a reference to the contents (and whether the table has been updated since the last time the application obtained its address or status) to closure, whose return value becomes the output.

Wraps CFE_TBL_GetAddress and CFE_TBL_ReleaseAddress.

pub fn load(&mut self, source: TblLoadSource<'_, T>) -> Result<(), Status>

Tries to load the table with data from source.

Wraps CFE_TBL_Load.

pub fn modified(&mut self) -> Result<(), Status>

Notifies Table Services that this application has modified the contents of this table.

Generally, applications using this crate won’t need to call this explicitly. The only mutable access to table contents this crate provides is through DumpOnlyTblHandle::get_mut, which calls CFE_TBL_Modified itself after any modification occurs.

Wraps CFE_TBL_Modified.

pub fn manage(&mut self) -> Result<bool, Status>

Performs the standard operations to maintain the table image.

Applications should call this periodically to process pending requests for updates, validation, or dumping to a buffer.

On success, returns whether a table update occurred.

Wraps CFE_TBL_Manage.

pub fn update(&mut self) -> Result<bool, Status>

Updates the contents of the table image, if an update is pending.

Applications should generally just use the manage method, which includes this bit of maintenance.

Returns whether there was, in fact, an update pending.

Wraps CFE_TBL_Update.

pub fn validate(&mut self) -> Result<bool, Status>

Validates the contents of the table image.

Applications should generally just use the manage method, which includes this bit of maintenance.

Returns whether there was, in fact, a validation request pending.

Wraps CFE_TBL_Validate.

pub fn dump_to_buffer(&mut self) -> Result<(), Status>

Copies the contents of a dump-only table to a buffer.

Applications should generally just use the manage method, which includes this bit of maintenance.

Wraps CFE_TBL_DumpToBuffer.

pub fn status(&self) -> Result<Option<PendingAction>, Status>

Returns one of the pending actions required for the table, if any.

Wraps CFE_TBL_GetStatus.

pub fn notify_by_message( &mut self, msg_id: MsgId, function_code: FunctionCode, payload: u32 ) -> Result<(), Status>

Instructs Table Services to notify the calling application with a message when the table requires management.

The message will have message ID msg_id, function code function_code, and u32 payload payload.

Only the application that owns the table in question may successfully call this.

Wraps CFE_TBL_NotifyByMessage.

pub fn unregister(self) -> Result<(), Status>

Unregisters the table corresponding to this handle.

Note that you generally shouldn’t need to call this for a table that the current application registered, as cFE automatically unregisters all tables owned by an application when that application exits.

Unregistering can be useful for handles to tables registered by other apps, as it frees up resources, but this is handled with the Drop impl on SharedTblHandle.

Wraps CFE_TBL_Unregister.