n2o4

Module osal

OSAL APIs.

Modules

file

Types and methods for interacting with files.

fs

File system-level APIs.

socket

Types and methods for interacting with network sockets.

sync

Synchronization primitives.

task

Task-related APIs.

Structs

OSTime

An instant in time.

OSTimeInterval

A time interval.

ObjectId

An identifier for an object managed by OSAL.

ObjectTypeConvertError

Error when trying to convert an ObjectId to a more-specialized type.

OsalError

An error code, as returned by many OSAL API functions.

Constants

MAX_NAME_LEN

The maximum length of strings for names of many OSAL objects.

MAX_PATH_LEN

The maximum allowed length of an OSAL path name, including directory name, file name, and terminating NUL character.