n2o4::osal
Module file

Types and methods for interacting with files.

Structs

File

A file handle.

FileFlags

Flags for opening/creating a File.

FileStat

Information about a file or directory.

OwnedFile

A wrapper for File that automatically closes its file handle when dropped.

Enums

AccessMode

The access mode a file should be opened with.

SeekReference

The reference point for a seek offset.

Functions

cp

Copies the file at src to dest.

file_open_check

Determines whether the file filename is open within OSAL.

mv

Moves the file at src to dest.

remove

Removes the file at path from the file system.

rename

Changes the name of the file originally at src to dest.

stat

Obtains information about the file or directory at path.