Struct n2o4::osal::socket::EarlySocket

pub struct EarlySocket<D: SocketDomain, T: SocketType> { /* private fields */ }

A network socket that has been created, but has yet to be either connected to a peer or bound to a local port.

Wraps osal_id_t.

Implementations

impl<D: SocketDomain, T: SocketType> EarlySocket<D, T>

pub fn open() -> Result<Self, i32>

Opens a socket with the given domain and type.

To do anything useful with the socket, connect or bind needs to be called.

Wraps OS_SocketOpen.

pub fn connect( self, addr: &SockAddr<D>, timeout_ms: Option<u32> ) -> Result<Socket<D, T, Connected>, i32>

Connects a socket to a peer at the remote address addr.

Waits up to timeout_ms.min(i32::MAX) milliseconds for a successful connection, or indefinitely if timeout_ms is None.

Wraps OS_SocketConnect.

pub fn bind(self, addr: &SockAddr<D>) -> Result<Socket<D, T, Bound>, i32>

Binds the socket to the local address addr.

Wraps OS_SocketBind.

pub fn as_id(&self) -> ObjectId

Returns the ObjectId for the socket.

pub unsafe fn from_id(id: ObjectId) -> Self

Unconditionally creates an EarlySocket from an OSAL ID.

Safety

This function does no checking that the ID in question corresponds to a socket, much less one of the correct domain, type, or point in socket lifecycle.

It is the programmer’s responsibility to ensure that any OSAL ID passed to from_id corresponds to a socket with the correct socket domain, type, and state.

pub fn info(&self) -> Result<SocketProperties, i32>

If successful, returns information about the socket.

Wraps OS_SocketGetInfo.