Module n2o4::sys

Low-level, unsafe bindings to the C APIs of cFE and OSAL.

This is n2o4’s equivalent of a *-sys crate for the cFE and OSAL APIs. Generally, users of n2o4 will not use this module directly, preferring the safe wrappers provided in the rest of n2o4. However, this module is available if one needs to use something for which safe wrappers haven’t been made.

The definitions in this module are auto-generated by bindgen, so there’s no Rustdoc comments available. Check out the cFE user’s guide for details about types and functions.

Those identifiers starting with S_, X_, and SHIM_ are the result of hand-written C code required to make bindgen create bindings for certain constants and function; each is the same as the C constant or function without the prefix in its name.

Re-exports

pub use self::CFE_TIME_Compare as CFE_TIME_Compare_t;

pub use self::CFE_TIME_Compare as CFE_TIME_Compare_t;

pub use self::CFE_MSG_Type as CFE_MSG_Type_t;

pub use self::CFE_MSG_SegmentationFlag as CFE_MSG_SegmentationFlag_t;

pub use self::CFE_MSG_Endian as CFE_MSG_Endian_t;

pub use self::CFE_MSG_PlaybackFlag as CFE_MSG_PlaybackFlag_t;

pub use self::CFE_TBL_SrcEnum as CFE_TBL_SrcEnum_t;

Structs

CCSDS_ExtendedHeader

CCSDS_PrimaryHeader

CCSDS_SpacePacket_t

CFE_ES_AppInfo

CFE_ES_AppNameCmd

CFE_ES_AppNameCmd_Payload

CFE_ES_AppReloadCmd_Payload

CFE_ES_BlockStats

CFE_ES_CDSRegDumpRec

CFE_ES_DeleteCDSCmd

CFE_ES_DeleteCDSCmd_Payload

CFE_ES_DumpCDSRegistryCmd

CFE_ES_DumpCDSRegistryCmd_Payload

CFE_ES_FileNameCmd

CFE_ES_FileNameCmd_Payload

CFE_ES_HousekeepingTlm

CFE_ES_HousekeepingTlm_Payload

CFE_ES_MemPoolStats

CFE_ES_MemStatsTlm

CFE_ES_NoArgsCmd

CFE_ES_OneAppTlm

CFE_ES_OneAppTlm_Payload

CFE_ES_OverWriteSysLogCmd

CFE_ES_OverWriteSysLogCmd_Payload

CFE_ES_PoolStatsTlm_Payload

CFE_ES_ReloadAppCmd

CFE_ES_RestartCmd

CFE_ES_RestartCmd_Payload

CFE_ES_SendMemPoolStatsCmd

CFE_ES_SendMemPoolStatsCmd_Payload

CFE_ES_SetMaxPRCountCmd

CFE_ES_SetMaxPRCountCmd_Payload

CFE_ES_SetPerfFilterMaskCmd

CFE_ES_SetPerfFilterMaskCmd_Payload

CFE_ES_SetPerfTrigMaskCmd_Payload

CFE_ES_SetPerfTriggerMaskCmd

CFE_ES_StartApp

CFE_ES_StartAppCmd_Payload

CFE_ES_StartPerfCmd_Payload

CFE_ES_StartPerfDataCmd

CFE_ES_StopPerfCmd_Payload

CFE_ES_StopPerfDataCmd

CFE_ES_TaskInfo

CFE_EVS_AppDataCmd_Payload

CFE_EVS_AppNameBitMaskCmd

CFE_EVS_AppNameBitMaskCmd_Payload

CFE_EVS_AppNameCmd

CFE_EVS_AppNameCmd_Payload

CFE_EVS_AppNameEventIDCmd

CFE_EVS_AppNameEventIDCmd_Payload

CFE_EVS_AppNameEventIDMaskCmd

CFE_EVS_AppNameEventIDMaskCmd_Payload

CFE_EVS_AppTlmData

CFE_EVS_BinFilter

CFE_EVS_BitMaskCmd

CFE_EVS_BitMaskCmd_Payload

CFE_EVS_HousekeepingTlm

CFE_EVS_HousekeepingTlm_Payload

CFE_EVS_LogFileCmd_Payload

CFE_EVS_LongEventTlm

CFE_EVS_LongEventTlm_Payload

CFE_EVS_NoArgsCmd

CFE_EVS_PacketID

CFE_EVS_SetEventFormatCode_Payload

CFE_EVS_SetEventFormatModeCmd

CFE_EVS_SetLogModeCmd

CFE_EVS_SetLogMode_Payload

CFE_EVS_ShortEventTlm

CFE_EVS_ShortEventTlm_Payload

CFE_EVS_WriteAppDataFileCmd

CFE_EVS_WriteLogDataFileCmd

CFE_FS_FileWriteMetaData

CFE_FS_Header

CFE_MSG_CommandHeader

CFE_MSG_CommandSecondaryHeader_t

CFE_MSG_TelemetryHeader

CFE_MSG_TelemetrySecondaryHeader_t

CFE_SB_AllSubscriptionsTlm

CFE_SB_AllSubscriptionsTlm_Payload

CFE_SB_HousekeepingTlm

CFE_SB_HousekeepingTlm_Payload

CFE_SB_MsgMapFileEntry

CFE_SB_PipeDepthStats

CFE_SB_PipeInfoEntry

CFE_SB_Qos_t

CFE_SB_RouteCmd

CFE_SB_RouteCmd_Payload

CFE_SB_RoutingFileEntry

CFE_SB_SingleSubscriptionTlm

CFE_SB_SingleSubscriptionTlm_Payload

CFE_SB_StatsTlm

CFE_SB_StatsTlm_Payload

CFE_SB_SubEntries

CFE_SB_WriteFileInfoCmd

CFE_SB_WriteFileInfoCmd_Payload

CFE_TBL_AbortLoadCmd

CFE_TBL_AbortLoadCmd_Payload

CFE_TBL_ActivateCmd

CFE_TBL_ActivateCmd_Payload

CFE_TBL_DelCDSCmd_Payload

CFE_TBL_DeleteCDSCmd

CFE_TBL_DumpCmd

CFE_TBL_DumpCmd_Payload

CFE_TBL_DumpRegistryCmd

CFE_TBL_DumpRegistryCmd_Payload

CFE_TBL_FileDef

CFE_TBL_File_Hdr

CFE_TBL_HousekeepingTlm

CFE_TBL_HousekeepingTlm_Payload

CFE_TBL_Info

CFE_TBL_LoadCmd

CFE_TBL_LoadCmd_Payload

CFE_TBL_NoArgsCmd

CFE_TBL_NotifyCmd

CFE_TBL_NotifyCmd_Payload

CFE_TBL_SendRegistryCmd

CFE_TBL_SendRegistryCmd_Payload

CFE_TBL_TableRegistryTlm

CFE_TBL_TblRegPacket_Payload

CFE_TBL_ValidateCmd

CFE_TBL_ValidateCmd_Payload

CFE_TIME_DiagnosticTlm

CFE_TIME_DiagnosticTlm_Payload

CFE_TIME_HousekeepingTlm

CFE_TIME_HousekeepingTlm_Payload

CFE_TIME_LeapsCmd_Payload

CFE_TIME_NoArgsCmd

CFE_TIME_OneHzAdjustmentCmd

CFE_TIME_OneHzAdjustmentCmd_Payload

CFE_TIME_SetLeapSecondsCmd

CFE_TIME_SetSignalCmd

CFE_TIME_SetSourceCmd

CFE_TIME_SetStateCmd

CFE_TIME_SignalCmd_Payload

CFE_TIME_SourceCmd_Payload

CFE_TIME_StateCmd_Payload

CFE_TIME_SysTime

CFE_TIME_TimeCmd

CFE_TIME_TimeCmd_Payload

CFE_TIME_ToneDataCmd

CFE_TIME_ToneDataCmd_Payload

OS_FdSet

OS_SockAddr_t

OS_bin_sem_prop_t

OS_count_sem_prop_t

OS_file_prop_t

OS_heap_prop_t

OS_module_address_t

OS_module_prop_t

OS_mut_sem_prop_t

OS_queue_prop_t

OS_socket_prop_t

OS_statvfs_t

OS_task_prop_t

OS_time_t

OS_timebase_prop_t

OS_timer_prop_t

_bindgen_ty_3

os_dirent_t

os_fsinfo_t

os_fstat_t

Constants

CFE_BUILD_BASELINE

CFE_BUILD_NUMBER

CFE_ES_ALL_APPS_EID

CFE_ES_APP_TLM_MID

CFE_ES_AppState_CFE_ES_AppState_EARLY_INIT

CFE_ES_AppState_CFE_ES_AppState_LATE_INIT

CFE_ES_AppState_CFE_ES_AppState_MAX

CFE_ES_AppState_CFE_ES_AppState_RUNNING

CFE_ES_AppState_CFE_ES_AppState_STOPPED

CFE_ES_AppState_CFE_ES_AppState_UNDEFINED

CFE_ES_AppState_CFE_ES_AppState_WAITING

CFE_ES_AppType_CFE_ES_AppType_CORE

CFE_ES_AppType_CFE_ES_AppType_EXTERNAL

CFE_ES_AppType_CFE_ES_AppType_LIBRARY

CFE_ES_BOOT_ERR_EID

CFE_ES_BUILD_INF_EID

CFE_ES_CC1_ERR_EID

CFE_ES_CDS_DELETED_INFO_EID

CFE_ES_CDS_DELETE_ERR_EID

CFE_ES_CDS_DELETE_TBL_ERR_EID

CFE_ES_CDS_DUMP_ERR_EID

CFE_ES_CDS_NAME_ERR_EID

CFE_ES_CDS_OWNER_ACTIVE_EID

CFE_ES_CDS_REGISTER_ERR_EID

CFE_ES_CDS_REG_DUMP_INF_EID

CFE_ES_CLEAR_ER_LOG_CC

CFE_ES_CLEAR_SYSLOG_CC

CFE_ES_CMD_MID

CFE_ES_CREATING_CDS_DUMP_ERR_EID

CFE_ES_DELETE_CDS_CC

CFE_ES_DUMP_CDS_REGISTRY_CC

CFE_ES_ERLOG1_INF_EID

CFE_ES_ERLOG2_EID

CFE_ES_ERLOG2_ERR_EID

CFE_ES_ERLOG_PENDING_ERR_EID

CFE_ES_ERREXIT_APP_ERR_EID

CFE_ES_ERREXIT_APP_INF_EID

CFE_ES_ERR_SYSLOGMODE_EID

CFE_ES_EXIT_APP_ERR_EID

CFE_ES_EXIT_APP_INF_EID

CFE_ES_ExceptionAction_CFE_ES_ExceptionAction_PROC_RESTART

CFE_ES_ExceptionAction_CFE_ES_ExceptionAction_RESTART_APP

CFE_ES_FILEWRITE_ERR_EID

CFE_ES_HK_TLM_MID

CFE_ES_INITSTATS_INF_EID

CFE_ES_INIT_INF_EID

CFE_ES_INVALID_POOL_HANDLE_ERR_EID

CFE_ES_LEN_ERR_EID

CFE_ES_LogEntryType_CFE_ES_LogEntryType_APPLICATION

CFE_ES_LogEntryType_CFE_ES_LogEntryType_CORE

CFE_ES_LogMode_CFE_ES_LogMode_DISCARD

CFE_ES_LogMode_CFE_ES_LogMode_OVERWRITE

CFE_ES_MEMSTATS_TLM_MID

CFE_ES_MID_ERR_EID

CFE_ES_NOOP_CC

CFE_ES_NOOP_INF_EID

CFE_ES_NO_MUTEX

CFE_ES_ONE_APPID_ERR_EID

CFE_ES_ONE_APP_EID

CFE_ES_ONE_ERR_EID

CFE_ES_OSCREATE_ERR_EID

CFE_ES_OVER_WRITE_SYSLOG_CC

CFE_ES_PCR_ERR1_EID

CFE_ES_PCR_ERR2_EID

CFE_ES_PERF_DATAWRITTEN_EID

CFE_ES_PERF_FILTMSKCMD_EID

CFE_ES_PERF_FILTMSKERR_EID

CFE_ES_PERF_LOG_ERR_EID

CFE_ES_PERF_STARTCMD_EID

CFE_ES_PERF_STARTCMD_ERR_EID

CFE_ES_PERF_STARTCMD_TRIG_ERR_EID

CFE_ES_PERF_STOPCMD_EID

CFE_ES_PERF_STOPCMD_ERR2_EID

CFE_ES_PERF_TRIGMSKCMD_EID

CFE_ES_PERF_TRIGMSKERR_EID

CFE_ES_QUERY_ALL_CC

CFE_ES_QUERY_ALL_TASKS_CC

CFE_ES_QUERY_ONE_CC

CFE_ES_RELOAD_APP_CC

CFE_ES_RELOAD_APP_DBG_EID

CFE_ES_RELOAD_APP_ERR1_EID

CFE_ES_RELOAD_APP_ERR2_EID

CFE_ES_RELOAD_APP_ERR3_EID

CFE_ES_RELOAD_APP_ERR4_EID

CFE_ES_RELOAD_APP_INF_EID

CFE_ES_RESET_COUNTERS_CC

CFE_ES_RESET_INF_EID

CFE_ES_RESET_PR_COUNT_CC

CFE_ES_RESET_PR_COUNT_EID

CFE_ES_RESTART_APP_CC

CFE_ES_RESTART_APP_DBG_EID

CFE_ES_RESTART_APP_ERR1_EID

CFE_ES_RESTART_APP_ERR2_EID

CFE_ES_RESTART_APP_ERR3_EID

CFE_ES_RESTART_APP_ERR4_EID

CFE_ES_RESTART_APP_INF_EID

CFE_ES_RESTART_CC

CFE_ES_RunStatus_CFE_ES_RunStatus_APP_ERROR

CFE_ES_RunStatus_CFE_ES_RunStatus_APP_EXIT

CFE_ES_RunStatus_CFE_ES_RunStatus_APP_RUN

CFE_ES_RunStatus_CFE_ES_RunStatus_CORE_APP_INIT_ERROR

CFE_ES_RunStatus_CFE_ES_RunStatus_CORE_APP_RUNTIME_ERROR

CFE_ES_RunStatus_CFE_ES_RunStatus_MAX

CFE_ES_RunStatus_CFE_ES_RunStatus_SYS_DELETE

CFE_ES_RunStatus_CFE_ES_RunStatus_SYS_EXCEPTION

CFE_ES_RunStatus_CFE_ES_RunStatus_SYS_RELOAD

CFE_ES_RunStatus_CFE_ES_RunStatus_SYS_RESTART

CFE_ES_RunStatus_CFE_ES_RunStatus_UNDEFINED

CFE_ES_SEND_HK_MID

CFE_ES_SEND_MEM_POOL_STATS_CC

CFE_ES_SET_MAX_PR_COUNT_CC

CFE_ES_SET_MAX_PR_COUNT_EID

CFE_ES_SET_PERF_FILTER_MASK_CC

CFE_ES_SET_PERF_TRIGGER_MASK_CC

CFE_ES_START_APP_CC

CFE_ES_START_ERR_EID

CFE_ES_START_EXC_ACTION_ERR_EID

CFE_ES_START_INF_EID

CFE_ES_START_INVALID_ENTRY_POINT_ERR_EID

CFE_ES_START_INVALID_FILENAME_ERR_EID

CFE_ES_START_NULL_APP_NAME_ERR_EID

CFE_ES_START_PERF_DATA_CC

CFE_ES_START_PRIORITY_ERR_EID

CFE_ES_STOP_APP_CC

CFE_ES_STOP_DBG_EID

CFE_ES_STOP_ERR1_EID

CFE_ES_STOP_ERR2_EID

CFE_ES_STOP_ERR3_EID

CFE_ES_STOP_INF_EID

CFE_ES_STOP_PERF_DATA_CC

CFE_ES_SYSLOG1_INF_EID

CFE_ES_SYSLOG2_EID

CFE_ES_SYSLOG2_ERR_EID

CFE_ES_SYSLOGMODE_EID

CFE_ES_SystemState_CFE_ES_SystemState_APPS_INIT

CFE_ES_SystemState_CFE_ES_SystemState_CORE_READY

CFE_ES_SystemState_CFE_ES_SystemState_CORE_STARTUP

CFE_ES_SystemState_CFE_ES_SystemState_EARLY_INIT

CFE_ES_SystemState_CFE_ES_SystemState_MAX

CFE_ES_SystemState_CFE_ES_SystemState_OPERATIONAL

CFE_ES_SystemState_CFE_ES_SystemState_SHUTDOWN

CFE_ES_SystemState_CFE_ES_SystemState_UNDEFINED

CFE_ES_TASKINFO_EID

CFE_ES_TASKINFO_OSCREATE_ERR_EID

CFE_ES_TASKINFO_WRHDR_ERR_EID

CFE_ES_TASKINFO_WR_ERR_EID

CFE_ES_TASKWR_ERR_EID

CFE_ES_TLM_POOL_STATS_INFO_EID

CFE_ES_USE_MUTEX

CFE_ES_VERSION_INF_EID

CFE_ES_WRHDR_ERR_EID

CFE_ES_WRITE_CFE_HDR_ERR_EID

CFE_ES_WRITE_ER_LOG_CC

CFE_ES_WRITE_SYSLOG_CC

CFE_EVS_ADDFILTER_EID

CFE_EVS_ADD_EVENT_FILTER_CC

CFE_EVS_CLEAR_LOG_CC

CFE_EVS_CMD_MID

CFE_EVS_CRITICAL_BIT

CFE_EVS_DEBUG_BIT

CFE_EVS_DELETE_EVENT_FILTER_CC

CFE_EVS_DELFILTER_EID

CFE_EVS_DISABLE_APP_EVENTS_CC

CFE_EVS_DISABLE_APP_EVENT_TYPE_CC

CFE_EVS_DISABLE_EVENT_TYPE_CC

CFE_EVS_DISABLE_PORTS_CC

CFE_EVS_DISAPPENTTYPE_EID

CFE_EVS_DISAPPEVT_EID

CFE_EVS_DISEVTTYPE_EID

CFE_EVS_DISPORT_EID

CFE_EVS_ENAAPPEVTTYPE_EID

CFE_EVS_ENAAPPEVT_EID

CFE_EVS_ENABLE_APP_EVENTS_CC

CFE_EVS_ENABLE_APP_EVENT_TYPE_CC

CFE_EVS_ENABLE_EVENT_TYPE_CC

CFE_EVS_ENABLE_PORTS_CC

CFE_EVS_ENAEVTTYPE_EID

CFE_EVS_ENAPORT_EID

CFE_EVS_ERROR_BIT

CFE_EVS_ERR_APPNOREGS_EID

CFE_EVS_ERR_CC_EID

CFE_EVS_ERR_CRDATFILE_EID

CFE_EVS_ERR_CRLOGFILE_EID

CFE_EVS_ERR_EVTIDNOREGS_EID

CFE_EVS_ERR_ILLAPPIDRANGE_EID

CFE_EVS_ERR_ILLEGALFMTMOD_EID

CFE_EVS_ERR_INVALID_BITMASK_EID

CFE_EVS_ERR_LOGMODE_EID

CFE_EVS_ERR_MAXREGSFILTER_EID

CFE_EVS_ERR_MSGID_EID

CFE_EVS_ERR_NOAPPIDFOUND_EID

CFE_EVS_ERR_UNREGISTERED_EVS_APP

CFE_EVS_ERR_WRDATFILE_EID

CFE_EVS_ERR_WRLOGFILE_EID

CFE_EVS_EVERY_FOURTH_ONE

CFE_EVS_EVERY_OTHER_ONE

CFE_EVS_EVERY_OTHER_TWO

CFE_EVS_EVT_FILTERED_EID

CFE_EVS_EventFilter_CFE_EVS_EventFilter_BINARY

CFE_EVS_EventOutput_CFE_EVS_EventOutput_PORT1

CFE_EVS_EventOutput_CFE_EVS_EventOutput_PORT2

CFE_EVS_EventOutput_CFE_EVS_EventOutput_PORT3

CFE_EVS_EventOutput_CFE_EVS_EventOutput_PORT4

CFE_EVS_EventType_CFE_EVS_EventType_CRITICAL

CFE_EVS_EventType_CFE_EVS_EventType_DEBUG

CFE_EVS_EventType_CFE_EVS_EventType_ERROR

CFE_EVS_EventType_CFE_EVS_EventType_INFORMATION

CFE_EVS_FILTER_MAX_EID

CFE_EVS_FIRST_4_STOP

CFE_EVS_FIRST_8_STOP

CFE_EVS_FIRST_16_STOP

CFE_EVS_FIRST_32_STOP

CFE_EVS_FIRST_64_STOP

CFE_EVS_FIRST_ONE_STOP

CFE_EVS_FIRST_TWO_STOP

CFE_EVS_HK_TLM_MID

CFE_EVS_INFORMATION_BIT

CFE_EVS_LEN_ERR_EID

CFE_EVS_LOGMODE_EID

CFE_EVS_LONG_EVENT_MSG_MID

CFE_EVS_LogMode_CFE_EVS_LogMode_DISCARD

CFE_EVS_LogMode_CFE_EVS_LogMode_OVERWRITE

CFE_EVS_MsgFormat_CFE_EVS_MsgFormat_LONG

CFE_EVS_MsgFormat_CFE_EVS_MsgFormat_SHORT

CFE_EVS_NOOP_CC

CFE_EVS_NOOP_EID

CFE_EVS_NO_FILTER

CFE_EVS_PORT1_BIT

CFE_EVS_PORT2_BIT

CFE_EVS_PORT3_BIT

CFE_EVS_PORT4_BIT

CFE_EVS_RESET_ALL_FILTERS_CC

CFE_EVS_RESET_APP_COUNTER_CC

CFE_EVS_RESET_COUNTERS_CC

CFE_EVS_RESET_FILTER_CC

CFE_EVS_RSTALLFILTER_EID

CFE_EVS_RSTCNT_EID

CFE_EVS_RSTEVTCNT_EID

CFE_EVS_RSTFILTER_EID

CFE_EVS_SEND_HK_MID

CFE_EVS_SETEVTFMTMOD_EID

CFE_EVS_SETFILTERMSK_EID

CFE_EVS_SET_EVENT_FORMAT_MODE_CC

CFE_EVS_SET_FILTER_CC

CFE_EVS_SET_LOG_MODE_CC

CFE_EVS_SHORT_EVENT_MSG_MID

CFE_EVS_STARTUP_EID

CFE_EVS_WRDAT_EID

CFE_EVS_WRITE_APP_DATA_FILE_CC

CFE_EVS_WRITE_LOG_DATA_FILE_CC

CFE_EVS_WRLOG_EID

CFE_FS_FILE_CONTENT_ID

CFE_FS_FileCategory_t_CFE_FS_FileCategory_BINARY_DATA_DUMP

CFE_FS_FileCategory_t_CFE_FS_FileCategory_DYNAMIC_MODULE

CFE_FS_FileCategory_t_CFE_FS_FileCategory_MAX

CFE_FS_FileCategory_t_CFE_FS_FileCategory_SCRIPT

CFE_FS_FileCategory_t_CFE_FS_FileCategory_TEMP

CFE_FS_FileCategory_t_CFE_FS_FileCategory_TEXT_LOG

CFE_FS_FileCategory_t_CFE_FS_FileCategory_UNKNOWN

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_COMPLETE

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_CREATE_ERROR

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_HEADER_WRITE_ERROR

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_MAX

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_RECORD_WRITE_ERROR

CFE_FS_FileWriteEvent_t_CFE_FS_FileWriteEvent_UNDEFINED

CFE_FS_HDR_DESC_MAX_LEN

CFE_FS_SubType_CFE_FS_SubType_ES_CDS_REG

CFE_FS_SubType_CFE_FS_SubType_ES_ERLOG

CFE_FS_SubType_CFE_FS_SubType_ES_PERFDATA

CFE_FS_SubType_CFE_FS_SubType_ES_QUERYALL

CFE_FS_SubType_CFE_FS_SubType_ES_QUERYALLTASKS

CFE_FS_SubType_CFE_FS_SubType_ES_SYSLOG

CFE_FS_SubType_CFE_FS_SubType_EVS_APPDATA

CFE_FS_SubType_CFE_FS_SubType_EVS_EVENTLOG

CFE_FS_SubType_CFE_FS_SubType_SB_MAPDATA

CFE_FS_SubType_CFE_FS_SubType_SB_PIPEDATA

CFE_FS_SubType_CFE_FS_SubType_SB_ROUTEDATA

CFE_FS_SubType_CFE_FS_SubType_TBL_IMG

CFE_FS_SubType_CFE_FS_SubType_TBL_REG

CFE_MAJOR_VERSION

CFE_MINOR_VERSION

CFE_MISSION_ES_APP_TLM_MSG

CFE_MISSION_ES_CDS_MAX_FULL_NAME_LEN

CFE_MISSION_ES_CDS_MAX_NAME_LENGTH

CFE_MISSION_ES_CMD_MSG

CFE_MISSION_ES_CRC_8

CFE_MISSION_ES_CRC_16

CFE_MISSION_ES_CRC_32

CFE_MISSION_ES_DEFAULT_CRC

CFE_MISSION_ES_HK_TLM_MSG

CFE_MISSION_ES_MAIN_PERF_ID

CFE_MISSION_ES_MAX_APPLICATIONS

CFE_MISSION_ES_MEMSTATS_TLM_MSG

CFE_MISSION_ES_PERF_EXIT_BIT

CFE_MISSION_ES_PERF_MAX_IDS

CFE_MISSION_ES_POOL_MAX_BUCKETS

CFE_MISSION_ES_SEND_HK_MSG

CFE_MISSION_EVS_CMD_MSG

CFE_MISSION_EVS_HK_TLM_MSG

CFE_MISSION_EVS_LONG_EVENT_MSG_MSG

CFE_MISSION_EVS_MAIN_PERF_ID

CFE_MISSION_EVS_MAX_MESSAGE_LENGTH

CFE_MISSION_EVS_SEND_HK_MSG

CFE_MISSION_EVS_SHORT_EVENT_MSG_MSG

CFE_MISSION_MAX_API_LEN

CFE_MISSION_MAX_FILE_LEN

CFE_MISSION_MAX_PATH_LEN

CFE_MISSION_REV

CFE_MISSION_SB_ALLSUBS_TLM_MSG

CFE_MISSION_SB_CMD_MSG

CFE_MISSION_SB_HK_TLM_MSG

CFE_MISSION_SB_MAIN_PERF_ID

CFE_MISSION_SB_MAX_PIPES

CFE_MISSION_SB_MAX_SB_MSG_SIZE

CFE_MISSION_SB_MSG_LIM_PERF_ID

CFE_MISSION_SB_ONESUB_TLM_MSG

CFE_MISSION_SB_PIPE_OFLOW_PERF_ID

CFE_MISSION_SB_SEND_HK_MSG

CFE_MISSION_SB_STATS_TLM_MSG

CFE_MISSION_SB_SUB_RPT_CTRL_MSG

CFE_MISSION_TBL_CMD_MSG

CFE_MISSION_TBL_HK_TLM_MSG

CFE_MISSION_TBL_MAIN_PERF_ID

CFE_MISSION_TBL_MAX_NAME_LENGTH

CFE_MISSION_TBL_REG_TLM_MSG

CFE_MISSION_TBL_SEND_HK_MSG

CFE_MISSION_TEST_CMD_MSG

CFE_MISSION_TEST_HK_TLM_MSG

CFE_MISSION_TIME_1HZ_CMD_MSG

CFE_MISSION_TIME_AT_TONE_WAS

CFE_MISSION_TIME_AT_TONE_WILL_BE

CFE_MISSION_TIME_CFG_DEFAULT_TAI

CFE_MISSION_TIME_CFG_DEFAULT_UTC

CFE_MISSION_TIME_CFG_FAKE_TONE

CFE_MISSION_TIME_CMD_MSG

CFE_MISSION_TIME_DATA_CMD_MSG

CFE_MISSION_TIME_DEF_DELAY_SECS

CFE_MISSION_TIME_DEF_DELAY_SUBS

CFE_MISSION_TIME_DEF_LEAPS

CFE_MISSION_TIME_DEF_MET_SECS

CFE_MISSION_TIME_DEF_MET_SUBS

CFE_MISSION_TIME_DEF_STCF_SECS

CFE_MISSION_TIME_DEF_STCF_SUBS

CFE_MISSION_TIME_DIAG_TLM_MSG

CFE_MISSION_TIME_EPOCH_DAY

CFE_MISSION_TIME_EPOCH_HOUR

CFE_MISSION_TIME_EPOCH_MINUTE

CFE_MISSION_TIME_EPOCH_SECOND

CFE_MISSION_TIME_EPOCH_YEAR

CFE_MISSION_TIME_FS_FACTOR

CFE_MISSION_TIME_HK_TLM_MSG

CFE_MISSION_TIME_LOCAL1HZISR_PERF_ID

CFE_MISSION_TIME_LOCAL1HZTASK_PERF_ID

CFE_MISSION_TIME_MAIN_PERF_ID

CFE_MISSION_TIME_MAX_ELAPSED

CFE_MISSION_TIME_MIN_ELAPSED

CFE_MISSION_TIME_SENDMET_PERF_ID

CFE_MISSION_TIME_SEND_CMD_MSG

CFE_MISSION_TIME_SEND_HK_MSG

CFE_MISSION_TIME_TONE1HZISR_PERF_ID

CFE_MISSION_TIME_TONE1HZTASK_PERF_ID

CFE_MISSION_TIME_TONE_CMD_MSG

CFE_MSG_Endian_CFE_MSG_Endian_Big

CFE_MSG_Endian_CFE_MSG_Endian_Invalid

CFE_MSG_Endian_CFE_MSG_Endian_Little

CFE_MSG_PlaybackFlag_CFE_MSG_PlayFlag_Invalid

CFE_MSG_PlaybackFlag_CFE_MSG_PlayFlag_Original

CFE_MSG_PlaybackFlag_CFE_MSG_PlayFlag_Playback

CFE_MSG_SegmentationFlag_CFE_MSG_SegFlag_Continue

CFE_MSG_SegmentationFlag_CFE_MSG_SegFlag_First

CFE_MSG_SegmentationFlag_CFE_MSG_SegFlag_Invalid

CFE_MSG_SegmentationFlag_CFE_MSG_SegFlag_Last

CFE_MSG_SegmentationFlag_CFE_MSG_SegFlag_Unsegmented

CFE_MSG_Type_CFE_MSG_Type_Cmd

CFE_MSG_Type_CFE_MSG_Type_Invalid

CFE_MSG_Type_CFE_MSG_Type_Tlm

CFE_PLATFORM_CMD_MID_BASE

CFE_PLATFORM_CMD_MID_BASE_GLOB

CFE_PLATFORM_CORE_MAX_STARTUP_MSEC

CFE_PLATFORM_ES_APP_KILL_TIMEOUT

CFE_PLATFORM_ES_APP_SCAN_RATE

CFE_PLATFORM_ES_CDS_MAX_BLOCK_SIZE

CFE_PLATFORM_ES_CDS_MAX_NUM_ENTRIES

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_01

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_02

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_03

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_04

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_05

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_06

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_07

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_08

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_09

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_10

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_11

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_12

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_13

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_14

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_15

CFE_PLATFORM_ES_CDS_MEM_BLOCK_SIZE_16

CFE_PLATFORM_ES_CDS_SIZE

CFE_PLATFORM_ES_DEFAULT_APP_LOG_FILE

CFE_PLATFORM_ES_DEFAULT_CDS_REG_DUMP_FILE

CFE_PLATFORM_ES_DEFAULT_ER_LOG_FILE

CFE_PLATFORM_ES_DEFAULT_PERF_DUMP_FILENAME

CFE_PLATFORM_ES_DEFAULT_POR_SYSLOG_MODE

CFE_PLATFORM_ES_DEFAULT_PR_SYSLOG_MODE

CFE_PLATFORM_ES_DEFAULT_STACK_SIZE

CFE_PLATFORM_ES_DEFAULT_SYSLOG_FILE

CFE_PLATFORM_ES_DEFAULT_TASK_LOG_FILE

CFE_PLATFORM_ES_ER_LOG_ENTRIES

CFE_PLATFORM_ES_ER_LOG_MAX_CONTEXT_SIZE

CFE_PLATFORM_ES_MAX_APPLICATIONS

CFE_PLATFORM_ES_MAX_BLOCK_SIZE

CFE_PLATFORM_ES_MAX_GEN_COUNTERS

CFE_PLATFORM_ES_MAX_LIBRARIES

CFE_PLATFORM_ES_MAX_MEMORY_POOLS

CFE_PLATFORM_ES_MAX_PROCESSOR_RESETS

CFE_PLATFORM_ES_MEMPOOL_ALIGN_SIZE_MIN

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_01

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_02

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_03

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_04

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_05

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_06

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_07

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_08

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_09

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_10

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_11

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_12

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_13

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_14

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_15

CFE_PLATFORM_ES_MEM_BLOCK_SIZE_16

CFE_PLATFORM_ES_NONVOL_DISK_MOUNT_STRING

CFE_PLATFORM_ES_NONVOL_STARTUP_FILE

CFE_PLATFORM_ES_OBJECT_TABLE_SIZE

CFE_PLATFORM_ES_PERF_CHILD_MS_DELAY

CFE_PLATFORM_ES_PERF_CHILD_PRIORITY

CFE_PLATFORM_ES_PERF_CHILD_STACK_SIZE

CFE_PLATFORM_ES_PERF_DATA_BUFFER_SIZE

CFE_PLATFORM_ES_PERF_ENTRIES_BTWN_DLYS

CFE_PLATFORM_ES_PERF_FILTMASK_ALL

CFE_PLATFORM_ES_PERF_FILTMASK_INIT

CFE_PLATFORM_ES_PERF_FILTMASK_NONE

CFE_PLATFORM_ES_PERF_TRIGMASK_ALL

CFE_PLATFORM_ES_PERF_TRIGMASK_INIT

CFE_PLATFORM_ES_PERF_TRIGMASK_NONE

CFE_PLATFORM_ES_POOL_MAX_BUCKETS

CFE_PLATFORM_ES_RAM_DISK_MOUNT_STRING

CFE_PLATFORM_ES_RAM_DISK_NUM_SECTORS

CFE_PLATFORM_ES_RAM_DISK_PERCENT_RESERVED

CFE_PLATFORM_ES_RAM_DISK_SECTOR_SIZE

CFE_PLATFORM_ES_RESET_AREA_SIZE

CFE_PLATFORM_ES_STARTUP_SCRIPT_TIMEOUT_MSEC

CFE_PLATFORM_ES_STARTUP_SYNC_POLL_MSEC

CFE_PLATFORM_ES_START_TASK_PRIORITY

CFE_PLATFORM_ES_START_TASK_STACK_SIZE

CFE_PLATFORM_ES_SYSTEM_LOG_SIZE

CFE_PLATFORM_ES_USER_RESERVED_SIZE

CFE_PLATFORM_ES_VOLATILE_STARTUP_FILE

CFE_PLATFORM_EVS_DEFAULT_APP_DATA_FILE

CFE_PLATFORM_EVS_DEFAULT_LOG_FILE

CFE_PLATFORM_EVS_DEFAULT_LOG_MODE

CFE_PLATFORM_EVS_DEFAULT_TYPE_FLAG

CFE_PLATFORM_EVS_LOG_MAX

CFE_PLATFORM_EVS_MAX_EVENT_FILTERS

CFE_PLATFORM_EVS_PORT_DEFAULT

CFE_PLATFORM_EVS_START_TASK_PRIORITY

CFE_PLATFORM_EVS_START_TASK_STACK_SIZE

CFE_PLATFORM_SB_BUF_MEMORY_BYTES

CFE_PLATFORM_SB_DEFAULT_MAP_FILENAME

CFE_PLATFORM_SB_DEFAULT_MSG_LIMIT

CFE_PLATFORM_SB_DEFAULT_PIPE_FILENAME

CFE_PLATFORM_SB_DEFAULT_ROUTING_FILENAME

CFE_PLATFORM_SB_FILTERED_EVENT5

CFE_PLATFORM_SB_FILTERED_EVENT6

CFE_PLATFORM_SB_FILTERED_EVENT7

CFE_PLATFORM_SB_FILTERED_EVENT8

CFE_PLATFORM_SB_FILTER_MASK1

CFE_PLATFORM_SB_FILTER_MASK2

CFE_PLATFORM_SB_FILTER_MASK3

CFE_PLATFORM_SB_FILTER_MASK4

CFE_PLATFORM_SB_FILTER_MASK5

CFE_PLATFORM_SB_FILTER_MASK6

CFE_PLATFORM_SB_FILTER_MASK7

CFE_PLATFORM_SB_FILTER_MASK8

CFE_PLATFORM_SB_HIGHEST_VALID_MSGID

CFE_PLATFORM_SB_MAX_BLOCK_SIZE

CFE_PLATFORM_SB_MAX_DEST_PER_PKT

CFE_PLATFORM_SB_MAX_MSG_IDS

CFE_PLATFORM_SB_MAX_PIPES

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_01

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_02

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_03

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_04

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_05

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_06

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_07

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_08

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_09

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_10

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_11

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_12

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_13

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_14

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_15

CFE_PLATFORM_SB_MEM_BLOCK_SIZE_16

CFE_PLATFORM_SB_START_TASK_PRIORITY

CFE_PLATFORM_SB_START_TASK_STACK_SIZE

CFE_PLATFORM_TBL_BUF_MEMORY_BYTES

CFE_PLATFORM_TBL_DEFAULT_REG_DUMP_FILE

CFE_PLATFORM_TBL_MAX_CRITICAL_TABLES

CFE_PLATFORM_TBL_MAX_DBL_TABLE_SIZE

CFE_PLATFORM_TBL_MAX_NUM_HANDLES

CFE_PLATFORM_TBL_MAX_NUM_TABLES

CFE_PLATFORM_TBL_MAX_NUM_VALIDATIONS

CFE_PLATFORM_TBL_MAX_SIMULTANEOUS_LOADS

CFE_PLATFORM_TBL_MAX_SNGL_TABLE_SIZE

CFE_PLATFORM_TBL_START_TASK_PRIORITY

CFE_PLATFORM_TBL_START_TASK_STACK_SIZE

CFE_PLATFORM_TBL_VALID_PRID_1

CFE_PLATFORM_TBL_VALID_PRID_3

CFE_PLATFORM_TBL_VALID_PRID_4

CFE_PLATFORM_TBL_VALID_PRID_COUNT

CFE_PLATFORM_TBL_VALID_SCID_1

CFE_PLATFORM_TBL_VALID_SCID_COUNT

CFE_PLATFORM_TIME_1HZ_TASK_PRIORITY

CFE_PLATFORM_TIME_1HZ_TASK_STACK_SIZE

CFE_PLATFORM_TIME_CFG_CLIENT

CFE_PLATFORM_TIME_CFG_LATCH_FLY

CFE_PLATFORM_TIME_CFG_SERVER

CFE_PLATFORM_TIME_CFG_SIGNAL

CFE_PLATFORM_TIME_CFG_SOURCE

CFE_PLATFORM_TIME_CFG_SRC_GPS

CFE_PLATFORM_TIME_CFG_SRC_MET

CFE_PLATFORM_TIME_CFG_SRC_TIME

CFE_PLATFORM_TIME_CFG_START_FLY

CFE_PLATFORM_TIME_CFG_TONE_LIMIT

CFE_PLATFORM_TIME_CFG_VIRTUAL

CFE_PLATFORM_TIME_MAX_DELTA_SECS

CFE_PLATFORM_TIME_MAX_DELTA_SUBS

CFE_PLATFORM_TIME_MAX_LOCAL_SECS

CFE_PLATFORM_TIME_MAX_LOCAL_SUBS

CFE_PLATFORM_TIME_START_TASK_PRIORITY

CFE_PLATFORM_TIME_START_TASK_STACK_SIZE

CFE_PLATFORM_TIME_TONE_TASK_PRIORITY

CFE_PLATFORM_TIME_TONE_TASK_STACK_SIZE

CFE_PLATFORM_TLM_MID_BASE

CFE_PSP_ERROR

CFE_PSP_ERROR_ADDRESS_MISALIGNED

CFE_PSP_ERROR_NOT_IMPLEMENTED

CFE_PSP_ERROR_TIMEOUT

CFE_PSP_INVALID_INT_NUM

CFE_PSP_INVALID_MEM_ADDR

CFE_PSP_INVALID_MEM_ATTR

CFE_PSP_INVALID_MEM_RANGE

CFE_PSP_INVALID_MEM_SIZE

CFE_PSP_INVALID_MEM_TYPE

CFE_PSP_INVALID_MEM_WORDSIZE

CFE_PSP_INVALID_MODULE_ID

CFE_PSP_INVALID_MODULE_NAME

CFE_PSP_INVALID_POINTER

CFE_PSP_MEM_ANY

CFE_PSP_MEM_ATTR_READ

CFE_PSP_MEM_ATTR_READWRITE

CFE_PSP_MEM_ATTR_WRITE

CFE_PSP_MEM_EEPROM

CFE_PSP_MEM_INVALID

CFE_PSP_MEM_RAM

CFE_PSP_MEM_SIZE_BYTE

CFE_PSP_MEM_SIZE_DWORD

CFE_PSP_MEM_SIZE_WORD

CFE_PSP_NO_EXCEPTION_DATA

CFE_PSP_PANIC_CORE_APP

CFE_PSP_PANIC_GENERAL_FAILURE

CFE_PSP_PANIC_MEMORY_ALLOC

CFE_PSP_PANIC_NONVOL_DISK

CFE_PSP_PANIC_STARTUP

CFE_PSP_PANIC_STARTUP_SEM

CFE_PSP_PANIC_VOLATILE_DISK

CFE_PSP_RST_SUBTYPE_BANKSWITCH_RESET

CFE_PSP_RST_SUBTYPE_EXCEPTION

CFE_PSP_RST_SUBTYPE_HWDEBUG_RESET

CFE_PSP_RST_SUBTYPE_HW_SPECIAL_COMMAND

CFE_PSP_RST_SUBTYPE_HW_WATCHDOG

CFE_PSP_RST_SUBTYPE_MAX

CFE_PSP_RST_SUBTYPE_POWER_CYCLE

CFE_PSP_RST_SUBTYPE_PUSH_BUTTON

CFE_PSP_RST_SUBTYPE_RESET_COMMAND

CFE_PSP_RST_SUBTYPE_UNDEFINED_RESET

CFE_PSP_RST_TYPE_MAX

CFE_PSP_RST_TYPE_POWERON

CFE_PSP_RST_TYPE_PROCESSOR

CFE_PSP_SOFT_TIMEBASE_NAME

CFE_PSP_SUCCESS

CFE_RESOURCEID_MARK

CFE_REVISION

CFE_SB_ALLSUBS_TLM_MID

CFE_SB_BAD_CMD_CODE_EID

CFE_SB_BAD_MSGID_EID

CFE_SB_BAD_PIPEID_EID

CFE_SB_CMD0_RCVD_EID

CFE_SB_CMD1_RCVD_EID

CFE_SB_CMD_MID

CFE_SB_CR_PIPE_BAD_ARG_EID

CFE_SB_CR_PIPE_ERR_EID

CFE_SB_CR_PIPE_NAME_TAKEN_EID

CFE_SB_CR_PIPE_NO_FREE_EID

CFE_SB_DEL_PIPE_ERR1_EID

CFE_SB_DEL_PIPE_ERR2_EID

CFE_SB_DEST_BLK_ERR_EID

CFE_SB_DISABLE_ROUTE_CC

CFE_SB_DISABLE_SUB_REPORTING_CC

CFE_SB_DSBL_RTE1_EID

CFE_SB_DSBL_RTE2_EID

CFE_SB_DSBL_RTE3_EID

CFE_SB_DUP_SUBSCRIP_EID

CFE_SB_ENABLE_ROUTE_CC

CFE_SB_ENABLE_SUB_REPORTING_CC

CFE_SB_ENBL_RTE1_EID

CFE_SB_ENBL_RTE2_EID

CFE_SB_ENBL_RTE3_EID

CFE_SB_FILEWRITE_ERR_EID

CFE_SB_FULL_SUB_PKT_EID

CFE_SB_GETPIPEIDBYNAME_EID

CFE_SB_GETPIPEIDBYNAME_NAME_ERR_EID

CFE_SB_GETPIPEIDBYNAME_NULL_ERR_EID

CFE_SB_GETPIPENAME_EID

CFE_SB_GETPIPENAME_ID_ERR_EID

CFE_SB_GETPIPENAME_NULL_PTR_EID

CFE_SB_GETPIPEOPTS_EID

CFE_SB_GETPIPEOPTS_ID_ERR_EID

CFE_SB_GETPIPEOPTS_PTR_ERR_EID

CFE_SB_GET_BUF_ERR_EID

CFE_SB_HASHCOLLISION_EID

CFE_SB_HK_TLM_MID

CFE_SB_INIT_EID

CFE_SB_LEN_ERR_EID

CFE_SB_MAX_DESTS_MET_EID

CFE_SB_MAX_MSGS_MET_EID

CFE_SB_MAX_PIPES_MET_EID

CFE_SB_MSGID_LIM_ERR_EID

CFE_SB_MSG_TOO_BIG_EID

CFE_SB_NOOP_CC

CFE_SB_ONESUB_TLM_MID

CFE_SB_PART_SUB_PKT_EID

CFE_SB_PEND_FOREVER

CFE_SB_PIPEOPTS_IGNOREMINE

CFE_SB_PIPE_ADDED_EID

CFE_SB_PIPE_DELETED_EID

CFE_SB_POLL

CFE_SB_Q_FULL_ERR_EID

CFE_SB_Q_RD_ERR_EID

CFE_SB_Q_WR_ERR_EID

CFE_SB_QosPriority_CFE_SB_QosPriority_HIGH

CFE_SB_QosPriority_CFE_SB_QosPriority_LOW

CFE_SB_QosReliability_CFE_SB_QosReliability_HIGH

CFE_SB_QosReliability_CFE_SB_QosReliability_LOW

CFE_SB_RCV_BAD_ARG_EID

CFE_SB_RESET_COUNTERS_CC

CFE_SB_SEND_BAD_ARG_EID

CFE_SB_SEND_HK_MID

CFE_SB_SEND_INV_MSGID_EID

CFE_SB_SEND_NO_SUBS_EID

CFE_SB_SEND_PREV_SUBS_CC

CFE_SB_SEND_SB_STATS_CC

CFE_SB_SETPIPEOPTS_EID

CFE_SB_SETPIPEOPTS_ID_ERR_EID

CFE_SB_SETPIPEOPTS_OWNER_ERR_EID

CFE_SB_SND_RTG_EID

CFE_SB_SND_RTG_ERR1_EID

CFE_SB_SND_STATS_EID

CFE_SB_STATS_TLM_MID

CFE_SB_SUBSCRIPTION

CFE_SB_SUBSCRIPTION_RCVD_EID

CFE_SB_SUBSCRIPTION_REMOVED_EID

CFE_SB_SUBSCRIPTION_RPT_EID

CFE_SB_SUB_ARG_ERR_EID

CFE_SB_SUB_ENTRIES_PER_PKT

CFE_SB_SUB_INV_CALLER_EID

CFE_SB_SUB_INV_PIPE_EID

CFE_SB_SUB_RPT_CTRL_MID

CFE_SB_UNSUBSCRIPTION

CFE_SB_UNSUB_ARG_ERR_EID

CFE_SB_UNSUB_INV_CALLER_EID

CFE_SB_UNSUB_INV_PIPE_EID

CFE_SB_UNSUB_NO_SUBS_EID

CFE_SB_WRITE_MAP_INFO_CC

CFE_SB_WRITE_PIPE_INFO_CC

CFE_SB_WRITE_ROUTING_INFO_CC

CFE_TBL_ABORT_LOAD_CC

CFE_TBL_ACTIVATE_CC

CFE_TBL_ACTIVATE_DUMP_ONLY_ERR_EID

CFE_TBL_ACTIVATE_ERR_EID

CFE_TBL_ASSUMED_VALID_INF_EID

CFE_TBL_BufferSelect_CFE_TBL_BufferSelect_ACTIVE

CFE_TBL_BufferSelect_CFE_TBL_BufferSelect_INACTIVE

CFE_TBL_CC1_ERR_EID

CFE_TBL_CDS_DELETED_INFO_EID

CFE_TBL_CDS_DELETE_ERR_EID

CFE_TBL_CDS_NOT_FOUND_ERR_EID

CFE_TBL_CDS_OWNER_ACTIVE_ERR_EID

CFE_TBL_CMD_MID

CFE_TBL_CREATING_DUMP_FILE_ERR_EID

CFE_TBL_DELETE_CDS_CC

CFE_TBL_DUMP_CC

CFE_TBL_DUMP_PENDING_ERR_EID

CFE_TBL_DUMP_REGISTRY_CC

CFE_TBL_FAIL_HK_SEND_ERR_EID

CFE_TBL_FAIL_NOTIFY_SEND_ERR_EID

CFE_TBL_FILE_ACCESS_ERR_EID

CFE_TBL_FILE_INCOMPLETE_ERR_EID

CFE_TBL_FILE_LOADED_INF_EID

CFE_TBL_FILE_STD_HDR_ERR_EID

CFE_TBL_FILE_SUBTYPE_ERR_EID

CFE_TBL_FILE_TBL_HDR_ERR_EID

CFE_TBL_FILE_TOO_BIG_ERR_EID

CFE_TBL_FILE_TYPE_ERR_EID

CFE_TBL_HANDLE_ACCESS_ERR_EID

CFE_TBL_HK_TLM_MID

CFE_TBL_ILLEGAL_BUFF_PARAM_ERR_EID

CFE_TBL_INIT_INF_EID

CFE_TBL_INTERNAL_ERROR_ERR_EID

CFE_TBL_IN_REGISTRY_ERR_EID

CFE_TBL_LEN_ERR_EID

CFE_TBL_LOADING_A_DUMP_ONLY_ERR_EID

CFE_TBL_LOADING_PENDING_ERR_EID

CFE_TBL_LOAD_ABORT_ERR_EID

CFE_TBL_LOAD_ABORT_INF_EID

CFE_TBL_LOAD_CC

CFE_TBL_LOAD_EXCEEDS_SIZE_ERR_EID

CFE_TBL_LOAD_FILENAME_LONG_ERR_EID

CFE_TBL_LOAD_IN_PROGRESS_ERR_EID

CFE_TBL_LOAD_PEND_REQ_INF_EID

CFE_TBL_LOAD_SUCCESS_INF_EID

CFE_TBL_LOAD_TBLNAME_MISMATCH_ERR_EID

CFE_TBL_LOAD_TYPE_ERR_EID

CFE_TBL_LOAD_VAL_ERR_EID

CFE_TBL_MID_ERR_EID

CFE_TBL_NOOP_CC

CFE_TBL_NOOP_INF_EID

CFE_TBL_NOT_CRITICAL_TBL_ERR_EID

CFE_TBL_NOT_IN_CRIT_REG_ERR_EID

CFE_TBL_NO_INACTIVE_BUFFER_ERR_EID

CFE_TBL_NO_SUCH_TABLE_ERR_EID

CFE_TBL_NO_WORK_BUFFERS_ERR_EID

CFE_TBL_OPT_BUFFER_MSK

CFE_TBL_OPT_CRITICAL

CFE_TBL_OPT_CRITICAL_MSK

CFE_TBL_OPT_DBL_BUFFER

CFE_TBL_OPT_DEFAULT

CFE_TBL_OPT_DUMP_ONLY

CFE_TBL_OPT_LD_DMP_MSK

CFE_TBL_OPT_LOAD_DUMP

CFE_TBL_OPT_NOT_CRITICAL

CFE_TBL_OPT_NOT_USR_DEF

CFE_TBL_OPT_SNGL_BUFFER

CFE_TBL_OPT_USR_DEF_ADDR

CFE_TBL_OPT_USR_DEF_MSK

CFE_TBL_OVERWRITE_DUMP_INF_EID

CFE_TBL_OVERWRITE_REG_DUMP_INF_EID

CFE_TBL_PARTIAL_LOAD_ERR_EID

CFE_TBL_PROCESSOR_ID_ERR_EID

CFE_TBL_REGISTER_ERR_EID

CFE_TBL_REG_TLM_MID

CFE_TBL_RESET_COUNTERS_CC

CFE_TBL_RESET_INF_EID

CFE_TBL_SEND_HK_MID

CFE_TBL_SEND_REGISTRY_CC

CFE_TBL_SHARE_ERR_EID

CFE_TBL_SPACECRAFT_ID_ERR_EID

CFE_TBL_SrcEnum_CFE_TBL_SRC_ADDRESS

CFE_TBL_SrcEnum_CFE_TBL_SRC_FILE

CFE_TBL_TLM_REG_CMD_INF_EID

CFE_TBL_TOO_MANY_DUMPS_ERR_EID

CFE_TBL_TOO_MANY_VALIDATIONS_ERR_EID

CFE_TBL_UNREGISTER_ERR_EID

CFE_TBL_UNVALIDATED_ERR_EID

CFE_TBL_UPDATE_ERR_EID

CFE_TBL_UPDATE_SUCCESS_INF_EID

CFE_TBL_VALIDATE_CC

CFE_TBL_VALIDATION_ERR_EID

CFE_TBL_VALIDATION_INF_EID

CFE_TBL_VAL_REQ_MADE_INF_EID

CFE_TBL_WRITE_CFE_HDR_ERR_EID

CFE_TBL_WRITE_DUMP_INF_EID

CFE_TBL_WRITE_REG_DUMP_INF_EID

CFE_TBL_WRITE_TBL_HDR_ERR_EID

CFE_TBL_WRITE_TBL_IMG_ERR_EID

CFE_TBL_WRITE_TBL_REG_ERR_EID

CFE_TBL_ZERO_LENGTH_LOAD_ERR_EID

CFE_TEST_CMD_MID

CFE_TEST_HK_TLM_MID

CFE_TIME_1HZ_CFG_EID

CFE_TIME_1HZ_CMD_MID

CFE_TIME_1HZ_EID

CFE_TIME_ADD_1HZ_ADJUSTMENT_CC

CFE_TIME_ADD_ADJUST_CC

CFE_TIME_ADD_DELAY_CC

CFE_TIME_AdjustDirection_CFE_TIME_AdjustDirection_ADD

CFE_TIME_AdjustDirection_CFE_TIME_AdjustDirection_SUBTRACT

CFE_TIME_CC_ERR_EID

CFE_TIME_CMD_MID

CFE_TIME_ClockState_CFE_TIME_ClockState_FLYWHEEL

CFE_TIME_ClockState_CFE_TIME_ClockState_INVALID

CFE_TIME_ClockState_CFE_TIME_ClockState_VALID

CFE_TIME_Compare_CFE_TIME_A_GT_B

CFE_TIME_Compare_CFE_TIME_A_LT_B

CFE_TIME_Compare_CFE_TIME_EQUAL

CFE_TIME_DATA_CMD_MID

CFE_TIME_DELAY_CFG_EID

CFE_TIME_DELAY_EID

CFE_TIME_DELAY_ERR_EID

CFE_TIME_DELTA_CFG_EID

CFE_TIME_DELTA_EID

CFE_TIME_DELTA_ERR_EID

CFE_TIME_DIAG_EID

CFE_TIME_DIAG_TLM_MID

CFE_TIME_FLAG_ADD1HZ

CFE_TIME_FLAG_ADDADJ

CFE_TIME_FLAG_ADDTCL

CFE_TIME_FLAG_CLKSET

CFE_TIME_FLAG_CMDFLY

CFE_TIME_FLAG_FLYING

CFE_TIME_FLAG_GDTONE

CFE_TIME_FLAG_REFERR

CFE_TIME_FLAG_SERVER

CFE_TIME_FLAG_SIGPRI

CFE_TIME_FLAG_SRCINT

CFE_TIME_FLAG_SRVFLY

CFE_TIME_FLAG_UNUSED

CFE_TIME_FLY_OFF_EID

CFE_TIME_FLY_ON_EID

CFE_TIME_FlagBit_CFE_TIME_FlagBit_ADD1HZ

CFE_TIME_FlagBit_CFE_TIME_FlagBit_ADDADJ

CFE_TIME_FlagBit_CFE_TIME_FlagBit_ADDTCL

CFE_TIME_FlagBit_CFE_TIME_FlagBit_CLKSET

CFE_TIME_FlagBit_CFE_TIME_FlagBit_CMDFLY

CFE_TIME_FlagBit_CFE_TIME_FlagBit_FLYING

CFE_TIME_FlagBit_CFE_TIME_FlagBit_GDTONE

CFE_TIME_FlagBit_CFE_TIME_FlagBit_SERVER

CFE_TIME_FlagBit_CFE_TIME_FlagBit_SIGPRI

CFE_TIME_FlagBit_CFE_TIME_FlagBit_SRCINT

CFE_TIME_FlagBit_CFE_TIME_FlagBit_SRVFLY

CFE_TIME_FlywheelState_CFE_TIME_FlywheelState_IS_FLY

CFE_TIME_FlywheelState_CFE_TIME_FlywheelState_NO_FLY

CFE_TIME_HK_TLM_MID

CFE_TIME_ID_ERR_EID

CFE_TIME_INIT_EID

CFE_TIME_LEAPS_CFG_EID

CFE_TIME_LEAPS_EID

CFE_TIME_LEN_ERR_EID

CFE_TIME_MET_CFG_EID

CFE_TIME_MET_EID

CFE_TIME_MET_ERR_EID

CFE_TIME_NOOP_CC

CFE_TIME_NOOP_EID

CFE_TIME_PRINTED_STRING_SIZE

CFE_TIME_RESET_COUNTERS_CC

CFE_TIME_RESET_EID

CFE_TIME_SEND_CMD_MID

CFE_TIME_SEND_DIAGNOSTIC_TLM_CC

CFE_TIME_SEND_HK_MID

CFE_TIME_SET_LEAP_SECONDS_CC

CFE_TIME_SET_MET_CC

CFE_TIME_SET_SIGNAL_CC

CFE_TIME_SET_SOURCE_CC

CFE_TIME_SET_STATE_CC

CFE_TIME_SET_STCF_CC

CFE_TIME_SET_TIME_CC

CFE_TIME_SIGNAL_CFG_EID

CFE_TIME_SIGNAL_EID

CFE_TIME_SIGNAL_ERR_EID

CFE_TIME_SOURCE_CFG_EID

CFE_TIME_SOURCE_EID

CFE_TIME_SOURCE_ERR_EID

CFE_TIME_STATE_EID

CFE_TIME_STATE_ERR_EID

CFE_TIME_STCF_CFG_EID

CFE_TIME_STCF_EID

CFE_TIME_STCF_ERR_EID

CFE_TIME_SUB_1HZ_ADJUSTMENT_CC

CFE_TIME_SUB_ADJUST_CC

CFE_TIME_SUB_DELAY_CC

CFE_TIME_SetState_CFE_TIME_SetState_NOT_SET

CFE_TIME_SetState_CFE_TIME_SetState_WAS_SET

CFE_TIME_SourceSelect_CFE_TIME_SourceSelect_EXTERNAL

CFE_TIME_SourceSelect_CFE_TIME_SourceSelect_INTERNAL

CFE_TIME_TIME_CFG_EID

CFE_TIME_TIME_EID

CFE_TIME_TIME_ERR_EID

CFE_TIME_TONE_CMD_MID

CFE_TIME_ToneSignalSelect_CFE_TIME_ToneSignalSelect_PRIMARY

CFE_TIME_ToneSignalSelect_CFE_TIME_ToneSignalSelect_REDUNDANT

OSAL_API_VERSION

OS_BUFFER_MSG_DEPTH

OS_BUFFER_SIZE

OS_BUILD_BASELINE

OS_BUILD_NUMBER

OS_CHECK

OS_CHK_ONLY

OS_ERROR

OS_ERROR_ADDRESS_MISALIGNED

OS_ERROR_NAME_LENGTH

OS_ERROR_TIMEOUT

OS_ERR_BAD_ADDRESS

OS_ERR_FILE

OS_ERR_INCORRECT_OBJ_STATE

OS_ERR_INCORRECT_OBJ_TYPE

OS_ERR_INVALID_ARGUMENT

OS_ERR_INVALID_ID

OS_ERR_INVALID_PRIORITY

OS_ERR_INVALID_SIZE

OS_ERR_NAME_NOT_FOUND

OS_ERR_NAME_TAKEN

OS_ERR_NAME_TOO_LONG

OS_ERR_NOT_IMPLEMENTED

OS_ERR_NO_FREE_IDS

OS_ERR_OBJECT_IN_USE

OS_ERR_OPERATION_NOT_SUPPORTED

OS_ERR_OUTPUT_TOO_LARGE

OS_ERR_SEM_NOT_FULL

OS_ERR_STREAM_DISCONNECTED

OS_Event_t_OS_EVENT_MAX

OS_Event_t_OS_EVENT_RESERVED

OS_Event_t_OS_EVENT_RESOURCE_ALLOCATED

OS_Event_t_OS_EVENT_RESOURCE_CREATED

OS_Event_t_OS_EVENT_RESOURCE_DELETED

OS_Event_t_OS_EVENT_TASK_STARTUP

OS_FILESTAT_MODE_DIR

OS_FILESTAT_MODE_EXEC

OS_FILESTAT_MODE_READ

OS_FILESTAT_MODE_WRITE

OS_FP_ENABLED

OS_FS_DEV_NAME_LEN

OS_FS_ERR_DEVICE_NOT_FREE

OS_FS_ERR_DRIVE_NOT_CREATED

OS_FS_ERR_NAME_TOO_LONG

OS_FS_ERR_PATH_INVALID

OS_FS_ERR_PATH_TOO_LONG

OS_FS_PHYS_NAME_LEN

OS_FS_VOL_NAME_LEN

OS_INVALID_INT_NUM

OS_INVALID_POINTER

OS_INVALID_SEM_VALUE

OS_MAJOR_VERSION

OS_MAX_API_NAME

OS_MAX_BIN_SEMAPHORES

OS_MAX_CMD_LEN

OS_MAX_CONSOLES

OS_MAX_COUNT_SEMAPHORES

OS_MAX_FILE_NAME

OS_MAX_FILE_SYSTEMS

OS_MAX_LOCAL_PATH_LEN

OS_MAX_MODULES

OS_MAX_MUTEXES

OS_MAX_NUM_OPEN_DIRS

OS_MAX_NUM_OPEN_FILES

OS_MAX_PATH_LEN

OS_MAX_QUEUES

OS_MAX_SYM_LEN

OS_MAX_TASKS

OS_MAX_TASK_PRIORITY

OS_MAX_TIMEBASES

OS_MAX_TIMERS

OS_MINOR_VERSION

OS_MISSION_REV

OS_MODULE_FILE_EXTENSION

OS_MODULE_FLAG_GLOBAL_SYMBOLS

OS_MODULE_FLAG_LOCAL_SYMBOLS

OS_OBJECT_INDEX_MASK

OS_OBJECT_TYPE_OS_BINSEM

OS_OBJECT_TYPE_OS_CONSOLE

OS_OBJECT_TYPE_OS_COUNTSEM

OS_OBJECT_TYPE_OS_DIR

OS_OBJECT_TYPE_OS_FILESYS

OS_OBJECT_TYPE_OS_MODULE

OS_OBJECT_TYPE_OS_MUTEX

OS_OBJECT_TYPE_OS_QUEUE

OS_OBJECT_TYPE_OS_STREAM

OS_OBJECT_TYPE_OS_TASK

OS_OBJECT_TYPE_OS_TIMEBASE

OS_OBJECT_TYPE_OS_TIMECB

OS_OBJECT_TYPE_SHIFT

OS_OBJECT_TYPE_UNDEFINED

OS_OBJECT_TYPE_USER

OS_PEND

OS_PRINTF_CONSOLE_NAME

OS_QUEUE_EMPTY

OS_QUEUE_FULL

OS_QUEUE_ID_ERROR

OS_QUEUE_INVALID_SIZE

OS_QUEUE_MAX_DEPTH

OS_QUEUE_TIMEOUT

OS_READ_ONLY

OS_READ_WRITE

OS_REPAIR

OS_REVISION

OS_SEEK_CUR

OS_SEEK_END

OS_SEEK_SET

OS_SEM_EMPTY

OS_SEM_FAILURE

OS_SEM_FULL

OS_SEM_TIMEOUT

OS_SHELL_CMD_INPUT_FILE_NAME

OS_SOCKADDR_MAX_LEN

OS_SUCCESS

OS_SocketDomain_t_OS_SocketDomain_INET

OS_SocketDomain_t_OS_SocketDomain_INET6

OS_SocketDomain_t_OS_SocketDomain_INVALID

OS_SocketDomain_t_OS_SocketDomain_MAX

OS_SocketShutdownMode_t_OS_SocketShutdownMode_NONE

OS_SocketShutdownMode_t_OS_SocketShutdownMode_SHUT_READ

OS_SocketShutdownMode_t_OS_SocketShutdownMode_SHUT_READWRITE

OS_SocketShutdownMode_t_OS_SocketShutdownMode_SHUT_WRITE

OS_SocketType_t_OS_SocketType_DATAGRAM

OS_SocketType_t_OS_SocketType_INVALID

OS_SocketType_t_OS_SocketType_MAX

OS_SocketType_t_OS_SocketType_STREAM

OS_StreamState_t_OS_STREAM_STATE_BOUND

OS_StreamState_t_OS_STREAM_STATE_CONNECTED

OS_StreamState_t_OS_STREAM_STATE_READABLE

OS_StreamState_t_OS_STREAM_STATE_WRITABLE

OS_TIMER_ERR_INTERNAL

OS_TIMER_ERR_INVALID_ARGS

OS_TIMER_ERR_TIMER_ID

OS_TIMER_ERR_UNAVAILABLE

OS_TIME_TICKS_PER_MSEC

OS_TIME_TICKS_PER_SECOND

OS_TIME_TICKS_PER_USEC

OS_TIME_TICK_RESOLUTION_NS

OS_UTILITYTASK_PRIORITY

OS_UTILITYTASK_STACK_SIZE

OS_VERSION_CODENAME

OS_WRITE_ONLY

OS_file_flag_t_OS_FILE_FLAG_CREATE

OS_file_flag_t_OS_FILE_FLAG_NONE

OS_file_flag_t_OS_FILE_FLAG_TRUNCATE

S_CFE_ES_APP_CLEANUP_ERR

S_CFE_ES_BAD_ARGUMENT

S_CFE_ES_BIN_SEM_DELETE_ERR

S_CFE_ES_BUFFER_NOT_IN_POOL

S_CFE_ES_CDS_ACCESS_ERROR

S_CFE_ES_CDS_ALREADY_EXISTS

S_CFE_ES_CDS_BLOCK_CRC_ERR

S_CFE_ES_CDS_INSUFFICIENT_MEMORY

S_CFE_ES_CDS_INVALID

S_CFE_ES_CDS_INVALID_NAME

S_CFE_ES_CDS_INVALID_SIZE

S_CFE_ES_CDS_OWNER_ACTIVE_ERR

S_CFE_ES_CDS_WRONG_TYPE_ERR

S_CFE_ES_COUNT_SEM_DELETE_ERR

S_CFE_ES_ERR_APP_CREATE

S_CFE_ES_ERR_APP_REGISTER

S_CFE_ES_ERR_CHILD_TASK_CREATE

S_CFE_ES_ERR_CHILD_TASK_DELETE

S_CFE_ES_ERR_CHILD_TASK_DELETE_MAIN_TASK

S_CFE_ES_ERR_CHILD_TASK_REGISTER

S_CFE_ES_ERR_DUPLICATE_NAME

S_CFE_ES_ERR_LOAD_LIB

S_CFE_ES_ERR_MEM_BLOCK_SIZE

S_CFE_ES_ERR_NAME_NOT_FOUND

S_CFE_ES_ERR_RESOURCEID_NOT_VALID

S_CFE_ES_ERR_SYS_LOG_FULL

S_CFE_ES_ERR_SYS_LOG_TRUNCATED

S_CFE_ES_FILE_CLOSE_ERR

S_CFE_ES_FILE_IO_ERR

S_CFE_ES_LIB_ALREADY_LOADED

S_CFE_ES_MUT_SEM_DELETE_ERR

S_CFE_ES_NOT_IMPLEMENTED

S_CFE_ES_NO_RESOURCE_IDS_AVAILABLE

S_CFE_ES_OPERATION_TIMED_OUT

S_CFE_ES_POOL_BLOCK_INVALID

S_CFE_ES_QUEUE_DELETE_ERR

S_CFE_ES_RST_ACCESS_ERR

S_CFE_ES_TASK_DELETE_ERR

S_CFE_ES_TIMER_DELETE_ERR

S_CFE_EVENTS_SERVICE

S_CFE_EVS_APP_FILTER_OVERLOAD

S_CFE_EVS_APP_ILLEGAL_APP_ID

S_CFE_EVS_APP_NOT_REGISTERED

S_CFE_EVS_EVT_NOT_REGISTERED

S_CFE_EVS_FILE_WRITE_ERROR

S_CFE_EVS_INVALID_PARAMETER

S_CFE_EVS_NOT_IMPLEMENTED

S_CFE_EVS_RESET_AREA_POINTER

S_CFE_EVS_UNKNOWN_FILTER

S_CFE_EXECUTIVE_SERVICE

S_CFE_FILE_SERVICE

S_CFE_FS_BAD_ARGUMENT

S_CFE_FS_FNAME_TOO_LONG

S_CFE_FS_INVALID_PATH

S_CFE_FS_NOT_IMPLEMENTED

S_CFE_GENERIC_SERVICE

S_CFE_MSG_WRONG_MSG_TYPE

S_CFE_SB_BAD_ARGUMENT

S_CFE_SB_BUFFER_INVALID

S_CFE_SB_BUF_ALOC_ERR

S_CFE_SB_INTERNAL_ERR

S_CFE_SB_MAX_DESTS_MET

S_CFE_SB_MAX_MSGS_MET

S_CFE_SB_MAX_PIPES_MET

S_CFE_SB_MSG_TOO_BIG

S_CFE_SB_NOT_IMPLEMENTED

S_CFE_SB_NO_MESSAGE

S_CFE_SB_PIPE_CR_ERR

S_CFE_SB_PIPE_RD_ERR

S_CFE_SB_TIME_OUT

S_CFE_SB_WRONG_MSG_TYPE

S_CFE_SERVICE_BITMASK

S_CFE_SEVERITY_BITMASK

S_CFE_SEVERITY_ERROR

S_CFE_SEVERITY_INFO

S_CFE_SEVERITY_SUCCESS

S_CFE_SOFTWARE_BUS_SERVICE

S_CFE_STATUS_BAD_COMMAND_CODE

S_CFE_STATUS_EXTERNAL_RESOURCE_FAIL

S_CFE_STATUS_NOT_IMPLEMENTED

S_CFE_STATUS_NO_COUNTER_INCREMENT

S_CFE_STATUS_REQUEST_ALREADY_PENDING

S_CFE_STATUS_UNKNOWN_MSG_ID

S_CFE_STATUS_WRONG_MSG_LENGTH

S_CFE_SUCCESS

S_CFE_TABLE_SERVICE

S_CFE_TBL_BAD_ARGUMENT

S_CFE_TBL_ERR_ACCESS

S_CFE_TBL_ERR_BAD_CONTENT_ID

S_CFE_TBL_ERR_BAD_PROCESSOR_ID

S_CFE_TBL_ERR_BAD_SPACECRAFT_ID

S_CFE_TBL_ERR_BAD_SUBTYPE_ID

S_CFE_TBL_ERR_DUMP_ONLY

S_CFE_TBL_ERR_DUPLICATE_DIFF_SIZE

S_CFE_TBL_ERR_DUPLICATE_NOT_OWNED

S_CFE_TBL_ERR_FILENAME_TOO_LONG

S_CFE_TBL_ERR_FILE_FOR_WRONG_TABLE

S_CFE_TBL_ERR_FILE_SIZE_INCONSISTENT

S_CFE_TBL_ERR_FILE_TOO_LARGE

S_CFE_TBL_ERR_HANDLES_FULL

S_CFE_TBL_ERR_ILLEGAL_SRC_TYPE

S_CFE_TBL_ERR_INVALID_HANDLE

S_CFE_TBL_ERR_INVALID_NAME

S_CFE_TBL_ERR_INVALID_OPTIONS

S_CFE_TBL_ERR_INVALID_SIZE

S_CFE_TBL_ERR_LOAD_INCOMPLETE

S_CFE_TBL_ERR_LOAD_IN_PROGRESS

S_CFE_TBL_ERR_NEVER_LOADED

S_CFE_TBL_ERR_NO_ACCESS

S_CFE_TBL_ERR_NO_BUFFER_AVAIL

S_CFE_TBL_ERR_NO_STD_HEADER

S_CFE_TBL_ERR_NO_TBL_HEADER

S_CFE_TBL_ERR_PARTIAL_LOAD

S_CFE_TBL_ERR_REGISTRY_FULL

S_CFE_TBL_ERR_SHORT_FILE

S_CFE_TBL_ERR_UNREGISTERED

S_CFE_TBL_INFO_DUMP_PENDING

S_CFE_TBL_INFO_NO_UPDATE_PENDING

S_CFE_TBL_INFO_NO_VALIDATION_PENDING

S_CFE_TBL_INFO_RECOVERED_TBL

S_CFE_TBL_INFO_TABLE_LOCKED

S_CFE_TBL_INFO_UPDATED

S_CFE_TBL_INFO_UPDATE_PENDING

S_CFE_TBL_INFO_VALIDATION_PENDING

S_CFE_TBL_MESSAGE_ERROR

S_CFE_TBL_NOT_IMPLEMENTED

S_CFE_TBL_WARN_DUPLICATE

S_CFE_TBL_WARN_NOT_CRITICAL

S_CFE_TBL_WARN_PARTIAL_LOAD

S_CFE_TBL_WARN_SHORT_FILE

S_CFE_TIME_BAD_ARGUMENT

S_CFE_TIME_CALLBACK_NOT_REGISTERED

S_CFE_TIME_INTERNAL_ONLY

S_CFE_TIME_NOT_IMPLEMENTED

S_CFE_TIME_OUT_OF_RANGE

S_CFE_TIME_SERVICE

S_CFE_TIME_TOO_MANY_SYNCH_CALLBACKS

X_CFE_ES_APPID_UNDEFINED

X_CFE_ES_CDS_BAD_HANDLE

X_CFE_ES_COUNTERID_UNDEFINED

X_CFE_ES_LIBID_UNDEFINED

X_CFE_ES_MEMHANDLE_UNDEFINED

X_CFE_ES_TASKID_UNDEFINED

X_CFE_RESOURCEID_RESERVED

X_CFE_RESOURCEID_UNDEFINED

X_CFE_SB_DEFAULT_QOS_PRIORITY

X_CFE_SB_DEFAULT_QOS_RELIABILITY

X_CFE_SB_INVALID_MSG_ID

X_CFE_SB_MSGID_RESERVED

X_CFE_TBL_BAD_TABLE_HANDLE

X_OS_OBJECT_ID_UNDEFINED

Statics

X_CFE_ES_TASK_STACK_ALLOCATE

Functions

CFE_ES_AppID_ToIndex⚠

CFE_ES_BackgroundWakeup⚠

CFE_ES_CalculateCRC⚠

CFE_ES_CopyToCDS⚠

CFE_ES_CounterID_ToIndex⚠

CFE_ES_CreateChildTask⚠

CFE_ES_DeleteApp⚠

CFE_ES_DeleteChildTask⚠

CFE_ES_DeleteGenCounter⚠

CFE_ES_ExitApp⚠

CFE_ES_ExitChildTask⚠

CFE_ES_GetAppID⚠

CFE_ES_GetAppIDByName⚠

CFE_ES_GetAppInfo⚠

CFE_ES_GetAppName⚠

CFE_ES_GetCDSBlockIDByName⚠

CFE_ES_GetCDSBlockName⚠

CFE_ES_GetGenCount⚠

CFE_ES_GetGenCounterIDByName⚠

CFE_ES_GetGenCounterName⚠

CFE_ES_GetLibIDByName⚠

CFE_ES_GetLibInfo⚠

CFE_ES_GetLibName⚠

CFE_ES_GetMemPoolStats⚠

CFE_ES_GetModuleInfo⚠

CFE_ES_GetPoolBuf⚠

CFE_ES_GetPoolBufInfo⚠

CFE_ES_GetResetType⚠

CFE_ES_GetTaskID⚠

CFE_ES_GetTaskIDByName⚠

CFE_ES_GetTaskInfo⚠

CFE_ES_GetTaskName⚠

CFE_ES_IncrementGenCounter⚠

CFE_ES_IncrementTaskCounter⚠

CFE_ES_LibID_ToIndex⚠

CFE_ES_PerfLogAdd⚠

CFE_ES_PoolCreate⚠

CFE_ES_PoolCreateEx⚠

CFE_ES_PoolCreateNoSem⚠

CFE_ES_PoolDelete⚠

CFE_ES_ProcessAsyncEvent⚠

CFE_ES_PutPoolBuf⚠

CFE_ES_RegisterCDS⚠

CFE_ES_RegisterGenCounter⚠

CFE_ES_ReloadApp⚠

CFE_ES_ResetCFE⚠

CFE_ES_RestartApp⚠

CFE_ES_RestoreFromCDS⚠

CFE_ES_RunLoop⚠

CFE_ES_SetGenCount⚠

CFE_ES_TaskID_ToIndex⚠

CFE_ES_WaitForStartupSync⚠

CFE_ES_WaitForSystemState⚠

CFE_ES_WriteToSysLog⚠

CFE_EVS_Register⚠

CFE_EVS_ResetAllFilters⚠

CFE_EVS_ResetFilter⚠

CFE_EVS_SendEvent⚠

CFE_EVS_SendEventWithAppID⚠

CFE_EVS_SendTimedEvent⚠

CFE_FS_BackgroundFileDumpIsPending⚠

CFE_FS_BackgroundFileDumpRequest⚠

CFE_FS_ExtractFilenameFromPath⚠

CFE_FS_GetDefaultExtension⚠

CFE_FS_GetDefaultMountPoint⚠

CFE_FS_InitHeader⚠

CFE_FS_ParseInputFileName⚠

CFE_FS_ParseInputFileNameEx⚠

CFE_FS_ReadHeader⚠

CFE_FS_SetTimestamp⚠

CFE_FS_WriteHeader⚠

CFE_MSG_GenerateChecksum⚠

CFE_MSG_GetApId⚠

CFE_MSG_GetEDSVersion⚠

CFE_MSG_GetEndian⚠

CFE_MSG_GetFcnCode⚠

CFE_MSG_GetHasSecondaryHeader⚠

CFE_MSG_GetHeaderVersion⚠

CFE_MSG_GetMsgId⚠

CFE_MSG_GetMsgTime⚠

CFE_MSG_GetNextSequenceCount⚠

CFE_MSG_GetPlaybackFlag⚠

CFE_MSG_GetSegmentationFlag⚠

CFE_MSG_GetSequenceCount⚠

CFE_MSG_GetSize⚠

CFE_MSG_GetSubsystem⚠

CFE_MSG_GetSystem⚠

CFE_MSG_GetType⚠

CFE_MSG_GetTypeFromMsgId⚠

CFE_MSG_Init⚠

CFE_MSG_SetApId⚠

CFE_MSG_SetEDSVersion⚠

CFE_MSG_SetEndian⚠

CFE_MSG_SetFcnCode⚠

CFE_MSG_SetHasSecondaryHeader⚠

CFE_MSG_SetHeaderVersion⚠

CFE_MSG_SetMsgId⚠

CFE_MSG_SetMsgTime⚠

CFE_MSG_SetPlaybackFlag⚠

CFE_MSG_SetSegmentationFlag⚠

CFE_MSG_SetSequenceCount⚠

CFE_MSG_SetSize⚠

CFE_MSG_SetSubsystem⚠

CFE_MSG_SetSystem⚠

CFE_MSG_SetType⚠

CFE_MSG_ValidateChecksum⚠

CFE_PSP_AttachExceptions⚠

CFE_PSP_Decompress⚠

CFE_PSP_EepromPowerDown⚠

CFE_PSP_EepromPowerUp⚠

CFE_PSP_EepromWrite8⚠

CFE_PSP_EepromWrite16⚠

CFE_PSP_EepromWrite32⚠

CFE_PSP_EepromWriteDisable⚠

CFE_PSP_EepromWriteEnable⚠

CFE_PSP_Exception_CopyContext⚠

CFE_PSP_Exception_GetCount⚠

CFE_PSP_Exception_GetSummary⚠

CFE_PSP_FlushCaches⚠

CFE_PSP_GetBuildNumber⚠

CFE_PSP_GetCDSSize⚠

CFE_PSP_GetCFETextSegmentInfo⚠

CFE_PSP_GetKernelTextSegmentInfo⚠

CFE_PSP_GetProcessorId⚠

CFE_PSP_GetProcessorName⚠

CFE_PSP_GetResetArea⚠

CFE_PSP_GetRestartType⚠

CFE_PSP_GetSpacecraftId⚠

CFE_PSP_GetTime⚠

CFE_PSP_GetTimerLow32Rollover⚠

CFE_PSP_GetTimerTicksPerSecond⚠

CFE_PSP_GetUserReservedArea⚠

CFE_PSP_GetVersionCodeName⚠

CFE_PSP_GetVersionNumber⚠

CFE_PSP_GetVersionString⚠

CFE_PSP_GetVolatileDiskMem⚠

CFE_PSP_Get_Dec⚠

CFE_PSP_Get_Timebase⚠

CFE_PSP_Get_Timer_Tick⚠

CFE_PSP_InitSSR⚠

CFE_PSP_Main⚠

CFE_PSP_MemCpy⚠

CFE_PSP_MemRangeGet⚠

CFE_PSP_MemRangeSet⚠

CFE_PSP_MemRanges⚠

CFE_PSP_MemRead8⚠

CFE_PSP_MemRead16⚠

CFE_PSP_MemRead32⚠

CFE_PSP_MemSet⚠

CFE_PSP_MemValidateRange⚠

CFE_PSP_MemWrite8⚠

CFE_PSP_MemWrite16⚠

CFE_PSP_MemWrite32⚠

CFE_PSP_Panic⚠

CFE_PSP_PortRead8⚠

CFE_PSP_PortRead16⚠

CFE_PSP_PortRead32⚠

CFE_PSP_PortWrite8⚠

CFE_PSP_PortWrite16⚠

CFE_PSP_PortWrite32⚠

CFE_PSP_ReadFromCDS⚠

CFE_PSP_Restart⚠

CFE_PSP_SetDefaultExceptionEnvironment⚠

CFE_PSP_WatchdogDisable⚠

CFE_PSP_WatchdogEnable⚠

CFE_PSP_WatchdogGet⚠

CFE_PSP_WatchdogInit⚠

CFE_PSP_WatchdogService⚠

CFE_PSP_WatchdogSet⚠

CFE_PSP_WriteToCDS⚠

CFE_ResourceId_FindNext⚠

CFE_ResourceId_GetBase⚠

CFE_ResourceId_GetSerial⚠

CFE_ResourceId_ToIndex⚠

CFE_SB_AllocateMessageBuffer⚠

CFE_SB_CreatePipe⚠

CFE_SB_DeletePipe⚠

CFE_SB_GetPipeIdByName⚠

CFE_SB_GetPipeName⚠

CFE_SB_GetPipeOpts⚠

CFE_SB_GetUserData⚠

CFE_SB_GetUserDataLength⚠

CFE_SB_IsValidMsgId⚠

CFE_SB_MessageStringGet⚠

CFE_SB_MessageStringSet⚠

CFE_SB_PipeId_ToIndex⚠

CFE_SB_ReceiveBuffer⚠

CFE_SB_ReleaseMessageBuffer⚠

CFE_SB_SetPipeOpts⚠

CFE_SB_SetUserDataLength⚠

CFE_SB_Subscribe⚠

CFE_SB_SubscribeEx⚠

CFE_SB_SubscribeLocal⚠

CFE_SB_TimeStampMsg⚠

CFE_SB_TransmitBuffer⚠

CFE_SB_TransmitMsg⚠

CFE_SB_Unsubscribe⚠

CFE_SB_UnsubscribeLocal⚠

CFE_TBL_DumpToBuffer⚠

CFE_TBL_GetAddress⚠

CFE_TBL_GetAddresses⚠

CFE_TBL_GetInfo⚠

CFE_TBL_GetStatus⚠

CFE_TBL_Load⚠

CFE_TBL_Manage⚠

CFE_TBL_Modified⚠

CFE_TBL_NotifyByMessage⚠

CFE_TBL_Register⚠

CFE_TBL_ReleaseAddress⚠

CFE_TBL_ReleaseAddresses⚠

CFE_TBL_Share⚠

CFE_TBL_Unregister⚠

CFE_TBL_Update⚠

CFE_TBL_Validate⚠

CFE_TIME_Add⚠

CFE_TIME_Compare⚠

CFE_TIME_ExternalGPS⚠

CFE_TIME_ExternalMET⚠

CFE_TIME_ExternalTime⚠

CFE_TIME_ExternalTone⚠

CFE_TIME_GetClockInfo⚠

CFE_TIME_GetClockState⚠

CFE_TIME_GetLeapSeconds⚠

CFE_TIME_GetMET⚠

CFE_TIME_GetMETseconds⚠

CFE_TIME_GetMETsubsecs⚠

CFE_TIME_GetSTCF⚠

CFE_TIME_GetTAI⚠

CFE_TIME_GetTime⚠

CFE_TIME_GetUTC⚠

CFE_TIME_Local1HzISR⚠

CFE_TIME_MET2SCTime⚠

CFE_TIME_Micro2SubSecs⚠

CFE_TIME_Print⚠

CFE_TIME_RegisterSynchCallback⚠

CFE_TIME_Sub2MicroSecs⚠

CFE_TIME_Subtract⚠

CFE_TIME_UnregisterSynchCallback⚠

OS_API_Init⚠

OS_API_Teardown⚠

OS_ApplicationExit⚠

OS_ApplicationShutdown⚠

OS_Application_Run⚠

OS_Application_Startup⚠

OS_BinSemCreate⚠

OS_BinSemDelete⚠

OS_BinSemFlush⚠

OS_BinSemGetIdByName⚠

OS_BinSemGetInfo⚠

OS_BinSemGive⚠

OS_BinSemTake⚠

OS_BinSemTimedWait⚠

OS_CloseAllFiles⚠

OS_CloseFileByName⚠

OS_ConvertToArrayIndex⚠

OS_CountSemCreate⚠

OS_CountSemDelete⚠

OS_CountSemGetIdByName⚠

OS_CountSemGetInfo⚠

OS_CountSemGive⚠

OS_CountSemTake⚠

OS_CountSemTimedWait⚠

OS_DeleteAllObjects⚠

OS_DirectoryClose⚠

OS_DirectoryOpen⚠

OS_DirectoryRead⚠

OS_DirectoryRewind⚠

OS_FDGetInfo⚠

OS_FS_GetPhysDriveName⚠

OS_FileOpenCheck⚠

OS_FileSysAddFixedMap⚠

OS_FileSysStatVolume⚠

OS_ForEachObject⚠

OS_ForEachObjectOfType⚠

OS_GetBuildNumber⚠

OS_GetErrorName⚠

OS_GetFsInfo⚠

OS_GetLocalTime⚠

OS_GetResourceName⚠

OS_GetVersionCodeName⚠

OS_GetVersionNumber⚠

OS_GetVersionString⚠

OS_HeapGetInfo⚠

OS_IdentifyObject⚠

OS_IdleLoop⚠

OS_ModuleInfo⚠

OS_ModuleLoad⚠

OS_ModuleSymbolLookup⚠

OS_ModuleUnload⚠

OS_MutSemCreate⚠

OS_MutSemDelete⚠

OS_MutSemGetIdByName⚠

OS_MutSemGetInfo⚠

OS_MutSemGive⚠

OS_MutSemTake⚠

OS_NetworkGetHostName⚠

OS_NetworkGetID⚠

OS_ObjectIdToArrayIndex⚠

OS_OpenCreate⚠

OS_QueueCreate⚠

OS_QueueDelete⚠

OS_QueueGet⚠

OS_QueueGetIdByName⚠

OS_QueueGetInfo⚠

OS_QueuePut⚠

OS_RegisterEventHandler⚠

OS_SelectFdAdd⚠

OS_SelectFdClear⚠

OS_SelectFdIsSet⚠

OS_SelectFdZero⚠

OS_SelectMultiple⚠

OS_SelectSingle⚠

OS_SetLocalTime⚠

OS_ShellOutputToFile⚠

OS_SocketAccept⚠

OS_SocketAddrFromString⚠

OS_SocketAddrGetPort⚠

OS_SocketAddrInit⚠

OS_SocketAddrSetPort⚠

OS_SocketAddrToString⚠

OS_SocketBind⚠

OS_SocketConnect⚠

OS_SocketGetIdByName⚠

OS_SocketGetInfo⚠

OS_SocketOpen⚠

OS_SocketRecvFrom⚠

OS_SocketSendTo⚠

OS_SocketShutdown⚠

OS_SymbolLookup⚠

OS_SymbolTableDump⚠

OS_TaskCreate⚠

OS_TaskDelay⚠

OS_TaskDelete⚠

OS_TaskExit⚠

OS_TaskFindIdBySystemData⚠

OS_TaskGetId⚠

OS_TaskGetIdByName⚠

OS_TaskGetInfo⚠

OS_TaskInstallDeleteHandler⚠

OS_TaskSetPriority⚠

OS_TimeBaseCreate⚠

OS_TimeBaseDelete⚠

OS_TimeBaseGetFreeRun⚠

OS_TimeBaseGetIdByName⚠

OS_TimeBaseGetInfo⚠

OS_TimeBaseSet⚠

OS_TimedRead⚠

OS_TimedWrite⚠

OS_TimerAdd⚠

OS_TimerCreate⚠

OS_TimerDelete⚠

OS_TimerGetIdByName⚠

OS_TimerGetInfo⚠

OS_TimerSet⚠

OS_TranslatePath⚠

OS_chkfs⚠

OS_chmod⚠

OS_close⚠

OS_cp⚠

OS_initfs⚠

OS_lseek⚠

OS_mkdir⚠

OS_mkfs⚠

OS_mount⚠

OS_mv⚠

OS_printf⚠

OS_printf_disable⚠

OS_printf_enable⚠

OS_read⚠

OS_remove⚠

OS_rename⚠

OS_rmdir⚠

OS_rmfs⚠

OS_stat⚠

OS_unmount⚠

OS_write⚠

SHIM_CFE_ResourceId_Equal⚠

SHIM_CFE_ResourceId_FromInteger⚠

SHIM_CFE_ResourceId_IsDefined⚠

SHIM_CFE_ResourceId_ToInteger⚠

SHIM_CFE_SB_MsgIdToValue⚠

SHIM_CFE_SB_MsgId_Equal⚠

SHIM_CFE_SB_ValueToMsgId⚠

SHIM_OS_ObjectIdDefined⚠

SHIM_OS_ObjectIdEqual⚠

SHIM_OS_ObjectIdFromInteger⚠

SHIM_OS_ObjectIdToInteger⚠

SHIM_OS_TimeAdd⚠

SHIM_OS_TimeAssembleFromMicroseconds⚠

SHIM_OS_TimeAssembleFromMilliseconds⚠

SHIM_OS_TimeAssembleFromNanoseconds⚠

SHIM_OS_TimeAssembleFromSubseconds⚠

SHIM_OS_TimeGetFractionalPart⚠

SHIM_OS_TimeGetMicrosecondsPart⚠

SHIM_OS_TimeGetMillisecondsPart⚠

SHIM_OS_TimeGetNanosecondsPart⚠

SHIM_OS_TimeGetSubsecondsPart⚠

SHIM_OS_TimeGetTotalMicroseconds⚠

SHIM_OS_TimeGetTotalMilliseconds⚠

SHIM_OS_TimeGetTotalNanoseconds⚠

SHIM_OS_TimeGetTotalSeconds⚠

SHIM_OS_TimeSubtract⚠

Type Aliases

CCSDS_ExtendedHeader_t

CCSDS_PrimaryHeader_t

CFE_ES_AppId_t

CFE_ES_AppInfo_t

CFE_ES_AppNameCmd_Payload_t

CFE_ES_AppNameCmd_t

CFE_ES_AppReloadCmd_Payload_t

CFE_ES_AppState

CFE_ES_AppState_Enum_t

CFE_ES_AppType

CFE_ES_AppType_Enum_t

CFE_ES_BlockStats_t

CFE_ES_CDSHandle_t

CFE_ES_CDSRegDumpRec_t

CFE_ES_ChildTaskMainFuncPtr_t

CFE_ES_ClearERLogCmd_t

CFE_ES_ClearSysLogCmd_t

CFE_ES_CounterId_t

CFE_ES_DeleteCDSCmd_Payload_t

CFE_ES_DeleteCDSCmd_t

CFE_ES_DumpCDSRegistryCmd_Payload_t

CFE_ES_DumpCDSRegistryCmd_t

CFE_ES_ExceptionAction

CFE_ES_ExceptionAction_Enum_t

CFE_ES_FileNameCmd_Payload_t

CFE_ES_FileNameCmd_t

CFE_ES_HousekeepingTlm_Payload_t

CFE_ES_HousekeepingTlm_t

CFE_ES_LibId_t

CFE_ES_LibraryEntryFuncPtr_t

CFE_ES_LogEntryType

CFE_ES_LogEntryType_Enum_t

CFE_ES_LogMode

CFE_ES_LogMode_Enum_t

CFE_ES_MemAddress_t

CFE_ES_MemHandle_t

CFE_ES_MemOffset_t

CFE_ES_MemPoolBuf_t

CFE_ES_MemPoolStats_t

CFE_ES_MemStatsTlm_t

CFE_ES_NoArgsCmd_t

CFE_ES_NoopCmd_t

CFE_ES_OneAppTlm_Payload_t

CFE_ES_OneAppTlm_t

CFE_ES_OverWriteSysLogCmd_Payload_t

CFE_ES_OverWriteSysLogCmd_t

CFE_ES_PoolAlign_t

CFE_ES_PoolStatsTlm_Payload_t

CFE_ES_QueryAllCmd_t

CFE_ES_QueryAllTasksCmd_t

CFE_ES_QueryOneCmd_t

CFE_ES_ReloadAppCmd_t

CFE_ES_ResetCountersCmd_t

CFE_ES_ResetPRCountCmd_t

CFE_ES_RestartAppCmd_t

CFE_ES_RestartCmd_Payload_t

CFE_ES_RestartCmd_t

CFE_ES_RunStatus

CFE_ES_RunStatus_Enum_t

CFE_ES_SendMemPoolStatsCmd_Payload_t

CFE_ES_SendMemPoolStatsCmd_t

CFE_ES_SetMaxPRCountCmd_Payload_t

CFE_ES_SetMaxPRCountCmd_t

CFE_ES_SetPerfFilterMaskCmd_Payload_t

CFE_ES_SetPerfFilterMaskCmd_t

CFE_ES_SetPerfTrigMaskCmd_Payload_t

CFE_ES_SetPerfTriggerMaskCmd_t

CFE_ES_StackPointer_t

CFE_ES_StartAppCmd_Payload_t

CFE_ES_StartAppCmd_t

CFE_ES_StartPerfCmd_Payload_t

CFE_ES_StartPerfDataCmd_t

CFE_ES_StopAppCmd_t

CFE_ES_StopPerfCmd_Payload_t

CFE_ES_StopPerfDataCmd_t

CFE_ES_SystemState

CFE_ES_SystemState_Enum_t

CFE_ES_TaskEntryFuncPtr_t

CFE_ES_TaskId_t

CFE_ES_TaskInfo_t

CFE_ES_TaskPriority_Atom_t

CFE_ES_WriteERLogCmd_t

CFE_ES_WriteSysLogCmd_t

CFE_EVS_AddEventFilterCmd_t

CFE_EVS_AppDataCmd_Payload_t

CFE_EVS_AppNameBitMaskCmd_Payload_t

CFE_EVS_AppNameBitMaskCmd_t

CFE_EVS_AppNameCmd_Payload_t

CFE_EVS_AppNameCmd_t

CFE_EVS_AppNameEventIDCmd_Payload_t

CFE_EVS_AppNameEventIDCmd_t

CFE_EVS_AppNameEventIDMaskCmd_Payload_t

CFE_EVS_AppNameEventIDMaskCmd_t

CFE_EVS_AppTlmData_t

CFE_EVS_BinFilter_t

CFE_EVS_BitMaskCmd_Payload_t

CFE_EVS_BitMaskCmd_t

CFE_EVS_ClearLogCmd_t

CFE_EVS_DeleteEventFilterCmd_t

CFE_EVS_DisableAppEventTypeCmd_t

CFE_EVS_DisableAppEventsCmd_t

CFE_EVS_DisableEventTypeCmd_t

CFE_EVS_DisablePortsCmd_t

CFE_EVS_EnableAppEventTypeCmd_t

CFE_EVS_EnableAppEventsCmd_t

CFE_EVS_EnableEventTypeCmd_t

CFE_EVS_EnablePortsCmd_t

CFE_EVS_EventFilter

CFE_EVS_EventFilter_Enum_t

CFE_EVS_EventOutput

CFE_EVS_EventOutput_Enum_t

CFE_EVS_EventType

CFE_EVS_EventType_Enum_t

CFE_EVS_HousekeepingTlm_Payload_t

CFE_EVS_HousekeepingTlm_t

CFE_EVS_LogFileCmd_Payload_t

CFE_EVS_LogMode

CFE_EVS_LogMode_Enum_t

CFE_EVS_LongEventTlm_Payload_t

CFE_EVS_LongEventTlm_t

CFE_EVS_MsgFormat

CFE_EVS_MsgFormat_Enum_t

CFE_EVS_NoArgsCmd_t

CFE_EVS_NoopCmd_t

CFE_EVS_PacketID_t

CFE_EVS_ResetAllFiltersCmd_t

CFE_EVS_ResetAppCounterCmd_t

CFE_EVS_ResetCountersCmd_t

CFE_EVS_ResetFilterCmd_t

CFE_EVS_SetEventFormatModeCmd_t

CFE_EVS_SetEventFormatMode_Payload_t

CFE_EVS_SetFilterCmd_t

CFE_EVS_SetLogModeCmd_t

CFE_EVS_SetLogMode_Payload_t

CFE_EVS_ShortEventTlm_Payload_t

CFE_EVS_ShortEventTlm_t

CFE_EVS_WriteAppDataFileCmd_t

CFE_EVS_WriteLogDataFileCmd_t

CFE_FS_FileCategory_t

CFE_FS_FileWriteEvent_t

CFE_FS_FileWriteGetData_t

CFE_FS_FileWriteMetaData_t

CFE_FS_FileWriteOnEvent_t

CFE_FS_Header_t

CFE_FS_SubType

CFE_FS_SubType_Enum_t

CFE_MSG_ApId_t

CFE_MSG_Checksum_t

CFE_MSG_CommandHeader_t

CFE_MSG_EDSVersion_t

CFE_MSG_Endian

CFE_MSG_FcnCode_t

CFE_MSG_HeaderVersion_t

CFE_MSG_Message_t

CFE_MSG_PlaybackFlag

CFE_MSG_SegmentationFlag

CFE_MSG_SequenceCount_t

CFE_MSG_Size_t

CFE_MSG_Subsystem_t

CFE_MSG_System_t

CFE_MSG_TelemetryHeader_t

CFE_MSG_Type

CFE_ResourceId_t

CFE_SB_AllSubscriptionsTlm_Payload_t

CFE_SB_AllSubscriptionsTlm_t

CFE_SB_Buffer_t

CFE_SB_DisableRouteCmd_t

CFE_SB_DisableSubReportingCmd_t

CFE_SB_EnableRouteCmd_t

CFE_SB_EnableSubReportingCmd_t

CFE_SB_HousekeepingTlm_Payload_t

CFE_SB_HousekeepingTlm_t

CFE_SB_MsgId_Atom_t

CFE_SB_MsgId_t

CFE_SB_MsgMapFileEntry_t

CFE_SB_NoopCmd_t

CFE_SB_PipeDepthStats_t

CFE_SB_PipeId_t

CFE_SB_PipeInfoEntry_t

CFE_SB_QosPriority

CFE_SB_QosPriority_Enum_t

CFE_SB_QosReliability

CFE_SB_QosReliability_Enum_t

CFE_SB_ResetCountersCmd_t

CFE_SB_RouteCmd_Payload_t

CFE_SB_RouteCmd_t

CFE_SB_RouteId_Atom_t

CFE_SB_RoutingFileEntry_t

CFE_SB_SendPrevSubsCmd_t

CFE_SB_SendSbStatsCmd_t

CFE_SB_SingleSubscriptionTlm_Payload_t

CFE_SB_SingleSubscriptionTlm_t

CFE_SB_StatsTlm_Payload_t

CFE_SB_StatsTlm_t

CFE_SB_SubEntries_t

CFE_SB_WriteFileInfoCmd_Payload_t

CFE_SB_WriteFileInfoCmd_t

CFE_SB_WriteMapInfoCmd_t

CFE_SB_WritePipeInfoCmd_t

CFE_SB_WriteRoutingInfoCmd_t

CFE_Status_t

CFE_TBL_AbortLoadCmd_Payload_t

CFE_TBL_AbortLoadCmd_t

CFE_TBL_ActivateCmd_Payload_t

CFE_TBL_ActivateCmd_t

CFE_TBL_BufferSelect

CFE_TBL_BufferSelect_Enum_t

CFE_TBL_CallbackFuncPtr_t

CFE_TBL_DelCDSCmd_Payload_t

CFE_TBL_DeleteCDSCmd_t

CFE_TBL_DumpCmd_Payload_t

CFE_TBL_DumpCmd_t

CFE_TBL_DumpRegistryCmd_Payload_t

CFE_TBL_DumpRegistryCmd_t

CFE_TBL_FileDef_t

CFE_TBL_File_Hdr_t

CFE_TBL_Handle_t

CFE_TBL_HousekeepingTlm_Payload_t

CFE_TBL_HousekeepingTlm_t

CFE_TBL_Info_t

CFE_TBL_LoadCmd_Payload_t

CFE_TBL_LoadCmd_t

CFE_TBL_NoArgsCmd_t

CFE_TBL_NoopCmd_t

CFE_TBL_NotifyCmd_Payload_t

CFE_TBL_NotifyCmd_t

CFE_TBL_ResetCountersCmd_t

CFE_TBL_SendRegistryCmd_Payload_t

CFE_TBL_SendRegistryCmd_t

CFE_TBL_SrcEnum

CFE_TBL_TableRegistryTlm_t

CFE_TBL_TblRegPacket_Payload_t

CFE_TBL_ValidateCmd_Payload_t

CFE_TBL_ValidateCmd_t

CFE_TIME_1HzCmd_t

CFE_TIME_Add1HZAdjustmentCmd_t

CFE_TIME_AddAdjustCmd_t

CFE_TIME_AddDelayCmd_t

CFE_TIME_AdjustDirection

CFE_TIME_AdjustDirection_Enum_t

CFE_TIME_ClockState

CFE_TIME_ClockState_Enum_t

CFE_TIME_Compare

CFE_TIME_DiagnosticTlm_Payload_t

CFE_TIME_DiagnosticTlm_t

CFE_TIME_FakeToneCmd_t

CFE_TIME_FlagBit

CFE_TIME_FlagBit_Enum_t

CFE_TIME_FlywheelState

CFE_TIME_FlywheelState_Enum_t

CFE_TIME_HousekeepingTlm_Payload_t

CFE_TIME_HousekeepingTlm_t

CFE_TIME_LeapsCmd_Payload_t

CFE_TIME_NoArgsCmd_t

CFE_TIME_NoopCmd_t

CFE_TIME_OneHzAdjustmentCmd_Payload_t

CFE_TIME_OneHzAdjustmentCmd_t

CFE_TIME_ResetCountersCmd_t

CFE_TIME_SendDiagnosticCmd_t

CFE_TIME_SetLeapSecondsCmd_t

CFE_TIME_SetMETCmd_t

CFE_TIME_SetSTCFCmd_t

CFE_TIME_SetSignalCmd_t

CFE_TIME_SetSourceCmd_t

CFE_TIME_SetState

CFE_TIME_SetStateCmd_t

CFE_TIME_SetState_Enum_t

CFE_TIME_SetTimeCmd_t

CFE_TIME_SignalCmd_Payload_t

CFE_TIME_SourceCmd_Payload_t

CFE_TIME_SourceSelect

CFE_TIME_SourceSelect_Enum_t

CFE_TIME_StateCmd_Payload_t

CFE_TIME_Sub1HZAdjustmentCmd_t

CFE_TIME_SubAdjustCmd_t

CFE_TIME_SubDelayCmd_t

CFE_TIME_SynchCallbackPtr_t

CFE_TIME_SysTime_t

CFE_TIME_TimeCmd_Payload_t

CFE_TIME_TimeCmd_t

CFE_TIME_ToneDataCmd_Payload_t

CFE_TIME_ToneDataCmd_t

CFE_TIME_ToneSignalCmd_t

CFE_TIME_ToneSignalSelect

CFE_TIME_ToneSignalSelect_Enum_t

OS_ArgCallback_t

OS_EventHandler_t

OS_Event_t

OS_SocketDomain_t

OS_SocketShutdownMode_t

OS_SocketType_t

OS_StreamState_t

OS_TimerCallback_t

OS_TimerSync_t

OS_file_flag_t

OS_static_symbol_record_t

__int16_t

__int32_t

__int64_t

__uint8_t

__uint16_t

__uint32_t

_bindgen_ty_1

_bindgen_ty_2

char

cpuaddr

int16

int32

int64

os_err_name_t

osal_blockcount_t

osal_id_t

osal_index_t

osal_objtype_t

osal_priority_t

osal_stackptr_t

osal_task

osal_task_entry

uint8

uint16

uint32

Unions

CFE_ES_PoolAlign

CFE_MSG_Message

CFE_SB_Msg

OS_SockAddrData_t