n2o4
Module utils

Utility types, functions, etc.

This module is for types that aren’t cFS-specific, but which turn out to be useful in APIs and not big enough to spin out into their own crates.

Structs

CStrBuf

An owned null-terminated C-compatible string of at most SIZE bytes (including null terminator).

NegativeI32

A wrapper for i32 that guarantees its value is always negative.

NotNegativeError

Error: an attempt was made to convert a non-negative value to a NegativeI32.

Enums

Unconstructable

A type for which no values can possibly exist.