Module n2o4::cfe::es

Executive Services system.

Structs

AppId

An identifier for cFE applications.

CdsHandle

A handle to a block in the Critical Data Store (CDS).

TaskFlags

Flags for task creation, as used by create_child_task.

TaskId

An identifier for cFE tasks.

TaskPriority

A task priority; used for task scheduling.

Enums

CdsRegisterResult

The possible varieties of successful outcome of CdsHandle::register/register_with/register_with_default.

ResetType

The type of cFE system reset desired in a call to reset_cfe.

RunStatus

The status (or requested status) of a cFE application.

SystemState

The current state of the overall cFS system.

Functions

create_child_task

Tries to create a new child task. If successful, runs function in the child task and returns the child task’s ID.

create_child_task_c

Tries to create a new child task. See create_child_task for details about the arguments.

delete_app

Stops a cFE application, then deletes it from the cFE application table.

exit_app

Exits from the current application.

exit_child_task

When called from a child task, causes the child task to exit and be deleted by cFE.

get_app_id

Returns (if successful) the application ID for the calling cFE application.

increment_task_counter

Increment ES’s internal counter for the current task.

perf_log_add

Logs an entry/exit marker for a specified ID for use by the Software Performance Analysis tool.

perf_log_entry

Shortcut for perf_log_add(marker, 0).

perf_log_exit

Shortcut for perf_log_add(marker, 1).

reload_app

Stops a cFE application, then loads and starts it using the specified file.

reset_cfe

Immediately resets the cFE core and all cFE applications.

restart_app

Restarts a single cFE application.

run_loop

Checks for exit requests from the cFE system and possibly makes a request for app shutdown to the cFE system.

wait_for_system_state

Waits for a minimum state of the overall cFS system, or a timeout (in milliseconds), whichever comes first.

write_to_syslog0

Writes a message to the cFE System Log using a format string and 0 format arguments.

write_to_syslog1

Writes a message to the cFE System Log using a format string and 1 format argument.

write_to_syslog2

Writes a message to the cFE System Log using a format string and 2 format arguments.

write_to_syslog3

Writes a message to the cFE System Log using a format string and 3 format arguments.

write_to_syslog4

Writes a message to the cFE System Log using a format string and 4 format arguments.

write_to_syslog5

Writes a message to the cFE System Log using a format string and 5 format arguments.

write_to_syslog6

Writes a message to the cFE System Log using a format string and 6 format arguments.

write_to_syslog7

Writes a message to the cFE System Log using a format string and 7 format arguments.

write_to_syslog8

Writes a message to the cFE System Log using a format string and 8 format arguments.

write_to_syslog_str

Writes the contents of a str to the cFE System Log.