n2o4::cfe::es
Function exit_child_task

pub fn exit_child_task() -> Result<Unconstructable, Status>

When called from a child task, causes the child task to exit and be deleted by cFE.

Unless an error occurs, this does not return.

Tasks created by create_child_task already call this automatically at the end of their execution, so functions passed to create_child_task do not need to manually call this function.

Wraps CFE_ES_ExitChildTask.