n2o4::cfe
Module evs

Event system.

Modules

bin_filter

Values intended for use in the Mask field of BinFilter.

Structs

BinFilter

Event-message filter definition for the binary filter scheme.

EventSender

A marker type to ensure you register before sending events.

Enums

EventType

The classification of an event message, analogous to the syslog severity level.

Traits

FilterScheme

A scheme for filtering event messages so that not all get recorded.

Functions

register

Registers the application with event services.