n2o4::cfe

Module sb

Software Bus system.

Structs

MsgId

An encoded message ID.

Pipe

A software bus pipe; an application needs one of these to receive messages.

Qos

Quality-of-service information for message subscriptions on the software bus. Currently unused by cFE.

Enums

QosPriority

Message priority for off-system routing. Currently unused by cFE.

QosReliability

Message transfer reliability for off-instance routing. Currently unused by cFE.

TimeOut

How long to wait for a new message if a pipe is empty.

Type Aliases

MsgId_Atom

The numeric value of a message ID.