Module n2o4::cfe::tbl

Table system.

Structs

DumpOnlyTblHandle

A handle to a dump-only table.

SharedTblHandle

A handle to a table registered by another application.

TableValidationFn

A wrapped version of a static fn to verify that a table (with contents of type T) is in a valid state.

TblHandle

A handle to a table.

TblInfo

The characteristics of, and statistical information for, a table.

TblOptions

Options available when registering a table using TblHandle::register.

Enums

PendingAction

A pending action for a table.

RegisterInfo

Alternative successful or partially-successful outcomes of TblHandle::register.

TblBuffering

Options regarding buffer use on table modifications.

TblCriticality

Options regarding whether a copy of the table is stored in the Critical Data Store (CDS).

TblLoadSource

A source of table-update data for TblHandle::load.

Traits

TableType

A convenience trait for referring to which types can be used as the contents of cFE tables.

Functions

info

Returns characteristics of, and information about, the table with name table_name.