Module n2o4::osal::socket

Types and methods for interacting with network sockets.

Structs

Bound

Marker type for sockets that have been bound to a local port.

Connected

Marker type for sockets that operate in a connection.

Datagram

Marker type for connectionless, message-oriented sockets.

EarlySocket

A network socket that has been created, but has yet to be either connected to a peer or bound to a local port.

IPv4

Marker type for IPv4 addresses and sockets.

IPv6

Marker type for IPv6 addresses and sockets.

SockAddr

A (network address, port) pair for use by socket-related APIs.

Socket

A network socket that is ready to send/receive data with a single peer (Connected) or accept new connections from/exchange datagrams with many peers (Bound).

SocketProperties

Information about a Socket or EarlySocket.

Stream

Marker type for connection-oriented, stream-of-bytes sockets.

Enums

SocketShutdownMode

The possible shutdown modes for a stream connection.

Traits

SocketDomain

A marker trait for network domains.

SocketRole

A marker trait for the role a socket plays.

SocketType

A marker trait for socket types.