Struct n2o4::osal::socket::Socket

pub struct Socket<D: SocketDomain, T: SocketType, R: SocketRole> { /* private fields */ }

A network socket that is ready to send/receive data with a single peer (Connected) or accept new connections from/exchange datagrams with many peers (Bound).

A Socket is either (1) created from an EarlySocket using connect or bind or (2) returned from a successful call to accept.

Certain operations only make sense for a subset of the possible combinations of SocketDomain, SocketType, and SocketRole.

Wraps osal_id_t.

Implementations

impl<D: SocketDomain, T: SocketType, R: SocketRole> Socket<D, T, R>

pub fn as_id(&self) -> ObjectId

Returns the ObjectId for the socket.

pub unsafe fn from_id(id: ObjectId, exclusive: bool) -> Self

Unconditionally creates a Socket from an OSAL ID, presumed to be that of an OSAL-managed socket.

exclusive indicates whether the generated Socket is the only possessor of the OSAL ID that might close or close_mut the OSAL socket.

Safety

This function does no checking that the ID in question corresponds to a socket, much less one of the correct domain, type, or point in socket lifecycle.

It is the programmer’s responsibility to ensure that any OSAL ID passed to from_id corresponds to a socket with the correct socket domain, type, state, and exclusivity.

pub fn close(self) -> Result<(), OsalError>

If the Socket hasn’t been Cloned, closes the socket.

If the socket has been cloned, returns Err(OS_ERR_OBJECT_IN_USE).

Wraps OS_close.

pub unsafe fn close_mut(&mut self) -> Result<(), OsalError>

Closes the socket.

This variant is intended to be used in Drop impls only; typically you want to use close instead.

Wraps OS_close.

Safety

This releases the underlying OSAL ID without necessarily destroying all references to the Socket. Any use of this Socket (or other Socket referring to the same underlying OSAL socket) after calling close_mut on it has potentially undesirable results—notably, there’s the possibility of the OSAL ID being reused for a different socket, leading to unintended use of another OSAL socket. As such, callers must make sure this Socket (and anything else using the same OSAL ID) is never used after calling close_mut.

pub fn info(&self) -> Result<SocketProperties, OsalError>

If successful, returns information about the socket.

Wraps OS_SocketGetInfo.

impl<D: SocketDomain, T: SocketType> Socket<D, T, Connected>

pub fn read(&self, buf: &mut [u8]) -> Result<usize, OsalError>

Reads up to buf.len() bytes from the connection into buf.

Upon success, returns the number of bytes actually read into buf, or 0 if at the end of the stream.

Wraps OS_read.

pub fn write(&self, buf: &[u8]) -> Result<usize, OsalError>

Writes up to buf.len() bytes from buf to the connection.

Upon success, returns the number of bytes actually written.

Wraps OS_write.

impl<D: SocketDomain> Socket<D, Stream, Connected>

pub fn shutdown(&self, mode: SocketShutdownMode) -> Result<(), OsalError>

Gracefully shuts down one or both directions of a stream connection.

Wraps OS_SocketShutdown.

impl<D: SocketDomain> Socket<D, Datagram, Connected>

pub fn connect( &self, addr: &SockAddr<D>, timeout_ms: Option<u32> ) -> Result<(), OsalError>

Tries to change the socket’s remote endpoint to addr, waiting up to timeout_ms.min(i32::MAX) to complete the operation.

Wraps OS_SocketConnect.

impl<D: SocketDomain> Socket<D, Stream, Bound>

pub fn accept( &self, timeout_ms: Option<u32> ) -> Result<(Socket<D, Stream, Connected>, SockAddr<D>), OsalError>

Waits for and accepts the next incoming connection on the given listening socket.

Waits for up to timeout_ms.min(i32::MAX) milliseconds for a new connection (or indefinitely if timeout_ms is None).

On success, results a socket for the new connection and the address of the connection’s remote side.

Wraps OS_SocketAccept.

impl<D: SocketDomain, R: SocketRole> Socket<D, Datagram, R>

pub fn send( &self, buf: &[u8], remote_addr: &SockAddr<D> ) -> Result<usize, OsalError>

Sends a message from the datagram socket to remote_addr, using buf as the message contents.

On success, returns the number of bytes of buf that were actually sent.

Wraps OS_SocketSendTo.

impl<D: SocketDomain> Socket<D, Datagram, Bound>

pub fn recv( &self, buf: &mut [u8], timeout_ms: Option<u32> ) -> Result<(usize, SockAddr<D>), OsalError>

Reads a message from the bound datagram socket into buf.

Wait up to timeout_ms.min(i32::MAX) milliseconds for a message (or indefinitely if timeout_ms is None).

On success, returns the number of bytes written to buf and the address of the message sender.

Wraps OS_SocketRecvFrom.